% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{AddAttributeDouble}
\alias{AddAttributeDouble}
\title{Attaches a numeric attribute to the current HDF5 file.}
\usage{
AddAttributeDouble(Object, AttributeName, Value)
}
\arguments{
\item{Object}{HDF5 object (group or dataset) to attach attribute (max. 255
characters).}

\item{AttributeName}{Attribute name (max. 127 characters).}

\item{Value}{Attribute value (numeric type).}
}
\description{
\code{AddAttributeDouble} attaches a numeric attribute to the current HDF5 file.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}

