% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddAttributeInt}
\alias{AddAttributeInt}
\title{Attaches an integer attribute to the current HDF5 file.}
\usage{
AddAttributeInt(Object, AttributeName, Value)
}
\arguments{
\item{Object}{HDF5 object (group or dataset) to attach attribute (max. 255
characters).}

\item{AttributeName}{Attribute name (max. 127 characters).}

\item{Value}{Attribute value (integer type).}
}
\description{
\code{AddAttributeInt} attaches an integer attribute to the current HDF5 file.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeString}},
  \code{\link{AddLogEntry}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}
\concept{Data storage functions}
