% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddAttributeString}
\alias{AddAttributeString}
\title{Attaches a string attribute to the current HDF5 file.}
\usage{
AddAttributeString(Object, AttributeName, Value)
}
\arguments{
\item{Object}{HDF5 object (group or dataset) to attach attribute (max. 255
characters).}

\item{AttributeName}{Attribute name (max. 127 characters).}

\item{Value}{Attribute string value (max. 255 characters).}
}
\description{
\code{AddAttributeString} attaches a string attribute to the current HDF5 file.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}}, \code{\link{AddLogEntry}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}
