% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{AddLogEntry}
\alias{AddLogEntry}
\title{Adds an entry to the acquisition log.}
\usage{
AddLogEntry(LogEntryText)
}
\description{
\code{AddLogEntry} adds an entry to the acquisition log.
}
\details{
Note: specifying the log entry time (for adding entries in the past) is not
supported.

@param LogEntryText Log text (max. 255 characters).

@family Data storage functions
}

