% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddUserDataMultiRow}
\alias{AddUserDataMultiRow}
\title{Stores (asynchronous) user supplied data.}
\usage{
AddUserDataMultiRow(Location, NbrElements, NbrRows, Data,
  ElementDescription = NULL, CompressionLevel = 0L)
}
\arguments{
\item{Location}{Location of group in HDF5 file where the datasets are created.}

\item{NbrElements}{Number of elements to store (per call to this function),
maximum is 1048575.}

\item{NbrRows}{Number of rows to store per call to this function (each row
contains \code{NbrElements} entries), maximum is 2047.}

\item{Data}{Vector of length \code{NbrElements} containing the data to be
stored in dataset "Data".}

\item{ElementDescription}{Vector of length \code{NbrElements} containing the
text description of elements. If \code{ElementDescription} is \code{NULL}
the dataset "Info" is not created.}

\item{CompressionLevel}{ZLIB compression level (0-9) for dataset creation.
If the dataset at Location already exists this parameter has no effect.}
}
\description{
\code{AddUserDataMultiRow} stores user supplied data asynchronously to the TOF data
acquistion into the current data file. Creates datasets "Data" and "Info" at
\code{Location}.
}
\details{
Same as \code{AddUserData}, but adds argument \code{NbrRows} to add several
lines of user data at once.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddLogEntry}}, \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}
