% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AutoSetupDaqDevice}
\alias{AutoSetupDaqDevice}
\title{Auto setup routine for the DAQ device.}
\usage{
AutoSetupDaqDevice()
}
\description{
\code{AutoSetupDaqDevice} sets up AP240 or Ndigo5G DAQ device.
}
\details{
This function is only functional for AP240 and Ndigo5G hardware. For all
other setups it returns success immediately but does not do anything.
}
