% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ChangePeakTable}
\alias{ChangePeakTable}
\title{Changes the PeakTable and recomputes the PeakData.}
\usage{
ChangePeakTable(Filename, PeakPar, compressionLevel)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{PeakPar}{List with the new peak paramters \emph{label}, \emph{mass},
\emph{loMass} and \emph{hiMass}.}

\item{compressionLevel}{Compression Level (1..9).}
}
\description{
\code{ChangePeakTable} changes the entries in the dataset \code{/PeakData/PeakTable}
and recomputes \code{/PeakData/PeakData} accordingly.
}
\details{
Depending on data file size recomputing the peak data can take a long time.
}
