% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ChangePeakTableFromFile}
\alias{ChangePeakTableFromFile}
\title{Changes the PeakTable and recomputes the PeakData.}
\usage{
ChangePeakTableFromFile(Filename, massTable, compressionLevel)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{massTable}{Path/filename to mass table or HDF5 file.}

\item{compressionLevel}{Compression Level (1..9).}
}
\description{
\code{ChangePeakTableFromFile} changes the entries in the dataset \code{/PeakData/PeakTable}
and recomputes \code{/PeakData/PeakData} accordingly. This is a convenience
function for \code{\link{ChangePeakTable}}. Instead of specifiying directly
peak parameters the information is taken from a mass table (text file) or a
Tofwerk HDF5 file.
}
\details{
Depending on data file size  recomputing the peak data can take a long time.
}
