% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{CleanupDll}
\alias{CleanupDll}
\title{Deinitializes the TofDaqDll.dll.}
\usage{
CleanupDll()
}
\description{
\code{CleanupDll} deinitializes the TofDaqDll.dll (frees allocated memory,
releases mapped shared memory and closes open files). This function is
automatically called when the TofDaqR packages is unloaded.
}

