% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CloseAll}
\alias{CloseAll}
\title{Closes all open HDF5 files.}
\usage{
CloseAll()
}
\description{
\code{CloseAll} closes all open HDF5 files. It is a good idea to call this
function once before your program exits.
}
\examples{
\dontrun{
CloseAll()
}
}

