% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CloseH5}
\alias{CloseH5}
\title{Closes an open HDF5 file.}
\usage{
CloseH5(Filename)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}
}
\description{
\code{CloseH5} closes an open HDF5 file.
}
\details{
This function is called internally by all \code{Get..FromH5} functions, so
it is usually not necessary to call \code{CloseH5} explicitely.
}
\examples{
\dontrun{
CloseH5("path/to/file.h5")
}
}
