% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ConfigureForSingleIonMeasurement}
\alias{ConfigureForSingleIonMeasurement}
\title{Configures TofDaq recorder for single ion measurements.}
\usage{
ConfigureForSingleIonMeasurement()
}
\value{
List with \code{nbrBits} and \code{negativeSignal} to pass to SI config function.
}
\description{
\code{ConfigureForSingleIonMeasurement} configures TofDaq recorder for single
ion measurements.
}
\details{
The function returns \code{nbrBits} and \code{negativeSignal} that need to be passed to
the single ion setup function in the tool library (all other parameters can
be read directly from TofDaq). Note that it is the user's responsibility to
backup current recorder settings before calling this function (and to revert
to this sertting after the SI run).
}
