% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ContinueAcquisition}
\alias{ContinueAcquisition}
\title{Signals to the TofDaq recorder to continue an acquisition.}
\usage{
ContinueAcquisition()
}
\description{
\code{ContinueAcquisition} signals to the TofDaq recorder to continue an
acquisition.
}
