% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{DaqActive}
\alias{DaqActive}
\title{Checks if TofDaq recorder is currently acquiring data.}
\usage{
DaqActive()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{DaqActive} checks if TofDaq recorder is currently acquiring data.
}

