% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DeleteAttributeInH5}
\alias{DeleteAttributeInH5}
\title{Deletes an attribute.}
\usage{
DeleteAttributeInH5(filename, location, name)
}
\arguments{
\item{location}{Location of the group or dataset the attribute is deleted from.}

\item{name}{Attribute name.}

\item{Filename}{Path/filename of the HDF5 file.}
}
\description{
\code{DeleteAttributeInH5} deletes an attribute.
}
\details{
WARNING: no sanity checking is performed! You can delete attributes that
render the data file unusable.
}
