% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DioStartDelayActive}
\alias{DioStartDelayActive}
\title{Checks if TofDaq recorder has received a start signal.}
\usage{
DioStartDelayActive()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{DioStartDelayActive} checks if TofDaq recorder has received a start
signal and is waiting for DioStartDelay to pass.
}
\details{
Signals \code{TRUE} when a start signal was received and "DioStartDelay" is
active. Goes \code{FALSE} when start delay has expired. In combination with
"TwWaitingForDioStartSignal" can be used to indicate to the user what the
current experiment state is when digital start signal (and delay) is used.
}
