% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EvalResolution}
\alias{EvalResolution}
\title{Evaluates the fitted resolution function for given mass values.}
\usage{
EvalResolution(R0, m0, dm, mass)
}
\arguments{
\item{R0}{Nominal mass resolution.}

\item{m0}{Mass where R(m) = 0.5*R0.}

\item{dm}{Slope parameter.}

\item{mass}{Vector of mass values.}
}
\value{
Vector with resolution values.
}
\description{
\code{EvalResolution} evaluates the fitted resolution function for given
mass values.
}
\details{
An empirical function that describes the mass resolution of TOFs as a
function of mass is: R(m) = R0 - R0/(1 + exp((m -m0)/dm)), where
R0 is the nominal mass resolution, m0 is the mass at which the resolution
is R0/2 and dm is a slope parameter.
}
\seealso{
\code{\link{FitResolution}}
}
