% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwToolDll.R
\name{EvalSinglePeak}
\alias{EvalSinglePeak}
\title{Evaluates a peak fit.}
\usage{
EvalSinglePeak(xVals, blOffset = 0, blSlope = 0, amplitude, fwhmLo,
  fwhmHi = fwhmLo, peakPos, mu = 0)
}
\arguments{
\item{xVals}{x axis data}

\item{blOffset}{Baseline offset at first data point.}

\item{blSlope}{Slope of baseline.}

\item{amplitude}{Peak amplitude.}

\item{fwhmLo}{Left peak width (full width at half maximum).}

\item{fwhmHi}{Right peak width (full width at half maximum).}

\item{peakPos}{Peak position.}

\item{mu}{Gaussian/Lorentzian contribution.}
}
\value{
Vector with y axis data.
}
\description{
\code{EvalSinglePeak} calculates the y-axis values for a given set of peak
parameters.
}
\seealso{
Other Peak fitting functions: \code{\link{FitSinglePeak2}},
  \code{\link{FitSinglePeak}}
}

