% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EventList2TofSpec}
\alias{EventList2TofSpec}
\title{Converts events of an event list into a spectrum.}
\usage{
EventList2TofSpec(events, clockPeriod, sampleInterval, nbrSamples)
}
\arguments{
\item{events}{Event data, e.g. from \code{GetEventList...FromH5}.}

\item{clockPeriod}{Clock period (in s or ns), e.g. from
\code{GetFloatAttributeFromH5(filename, "FullSpectra", "ClockPeriod")}.}

\item{sampleInterval}{Sampling interval (in same units as \code{clockPeriod}),
e.g. from \code{\link{GetH5Descriptor}}.}

\item{nbrSamples}{Number of samples, e.g. from \code{\link{GetH5Descriptor}}.}
}
\value{
A vector containing the spectrum.
}
\description{
\code{EventList2TofSpec} converts events of an event list (read with
\code{GetEventList...FromH5} functions) into a spectrum.
}
\details{
Note: this function is not part of the TofDaq API, but is included in the
package for convenience.
}

