% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FindTpsIp}
\alias{FindTpsIp}
\title{Gets IP address of TPS2.}
\usage{
FindTpsIp(TpsSerial, timeout)
}
\arguments{
\item{TpsSerial}{Serial number of TPS2.}

\item{timeout}{Timeout in ms to wait for the UDP packet.}
}
\value{
String of IP address.
}
\description{
\code{FindTpsIp} listens for UDP packets that TPS2 broadcast and returns
the IP of the TPS2.
}
\details{
Note that executing this function makes your program to a UDP server and
Windows firewall (or other personal firewall software) will query for
permission.
}
\examples{
\dontrun{
FindTpsIp("910.33.0316", 500)
}
}
