% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitResolution}
\alias{FitResolution}
\title{Fits mass versus resolution values to an empirical function.}
\usage{
FitResolution(mass, resolution, R0, m0, dm)
}
\arguments{
\item{mass}{Vector of mass values.}

\item{resolution}{Vector of resolution values.}

\item{R0}{Initial guess value of nominal mass resolution R0.}

\item{m0}{Initial guess value of m0 (mass where R(m) = 0.5*R0).}

\item{dm}{Initial guess value of slope parameter dm.}
}
\value{
List with the fitted parameters R0, m0 and dm.
}
\description{
\code{FitResolution} fits mass versus resolution values to an empirical
function (see Details).
}
\details{
An empirical function that describes the mass resolution as a function of
mass is: R(m) = R0 - R0/(1 + exp((m -m0)/dm)), where R0 is the nominal mass
resolution, m0 is the mass at which the resolution is R0/2 and dm is a slope
parameter.

Reasonable initial guess values for R0, m0 and dm need to be provided.
}
\seealso{
\code{\link{EvalResolution}}
}
