% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitSinglePeak}
\alias{FitSinglePeak}
\title{Performs a peak fit.}
\usage{
FitSinglePeak(yVals, xVals, peakType = 0L, blOffset = 0, blSlope = 0,
  amplitude = 0, fwhmLo = 0, fwhmHi = 0, peakPos = 0, mu = 0)
}
\arguments{
\item{yVals}{y axis data.}

\item{xVals}{x axis data.}

\item{peakType}{peak model to use.}

\item{blOffset}{Initial value of baseline offset at first data point.}

\item{blSlope}{Initial value of slope of baseline.}

\item{amplitude}{Initial value of peak amplitude.}

\item{fwhmLo}{Initial value of left peak width (full width at half maximum).}

\item{fwhmHi}{Initial value of right peak width (full width at half maximum).}

\item{peakPos}{Initial value of peak position.}

\item{mu}{Initial value of Gaussian/Lorentzian contribution.}
}
\value{
List with peak parameters.
}
\description{
\code{FitSinglePeak} performs a peak fit. Gaussian, Lorentzian and
Pseudo-Voigt peak shapes are supported.
}
\details{
Peak parameters are optimized using a Levenberg-Marquardt algorithm as
implemented by the library \code{lmfit}. The following peak types are available:
\tabular{cccc}{
peak type index \tab peak function \tab symmetric \tab baseline \cr
-1 \tab Gaussian through highest point and its neighbours \tab yes \tab no \cr
0 \tab Gaussian \tab yes \tab no  \cr
1 \tab Lorentzian \tab yes \tab no \cr
2 \tab Pseudo-Voigt \tab yes \tab no \cr
3 \tab Gaussian \tab yes \tab yes \cr
4 \tab Lorentzian \tab yes \tab yes \cr
5 \tab Pseudo-Voigt \tab yes \tab yes \cr
6 \tab Gaussian \tab no \tab no \cr
7 \tab Lorentzian \tab no \tab no \cr
8 \tab Pseudo-Voigt \tab no \tab no \cr
9 \tab Gaussian \tab no \tab yes \cr
10 \tab Lorentzian \tab no \tab yes \cr
11 \tab Pseudo-Voigt \tab no \tab yes
} The function for peakType = -1 is not an actual peak fit but rather the
Gaussian function (symmetric, no baseline) defined by the maximum intensity
point and its neighbors. It is used to automatically generate guess values
for position, width and amplitude if all peak parameters are set to 0.
}
\seealso{
Other Peak fitting functions: \code{\link{EvalMultiPeak}},
  \code{\link{EvalSinglePeak}},
  \code{\link{FitSinglePeak2}}, \code{\link{MultiPeakFit}}
}
\concept{Peak fitting functions}
