% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitSinglePeak2}
\alias{FitSinglePeak2}
\title{Performs a peak fit (Initial values as vector).}
\usage{
FitSinglePeak2(yVals, xVals, peakType = 0L, param = as.numeric(c(0, 0,
  0, 0, 0, 0, 0)))
}
\arguments{
\item{yVals}{y axis data.}

\item{xVals}{x axis data.}

\item{peakType}{peak model to use.}

\item{param}{Vector of initial values (blOffset, blSlope, amplitude, fwhmLo,
fwhmHi, peakPos, mu).}
}
\value{
List with peak parameters.
}
\description{
\code{FitSinglePeak2} performs a peak fit. Gaussian, Lorentzian and
Pseudo-Voigt peak shapes are supported.
}
\details{
Same as \code{\link{FitSinglePeak}}, but takes the initial values of the
fit parameter as a vector argument instead of explicit parameters.
Peak parameters are optimized using a Levenberg-Marquardt algorithm as
implemented by the library \code{lmfit}. The following peak types are available:
\tabular{cccc}{
peak type index \tab peak function \tab symmetric \tab baseline \cr
-1 \tab Gaussian through highest point and its neighbours \tab yes \tab no \cr
0 \tab Gaussian \tab yes \tab no  \cr
1 \tab Lorentzian \tab yes \tab no \cr
2 \tab Pseudo-Voigt \tab yes \tab no \cr
3 \tab Gaussian \tab yes \tab yes \cr
4 \tab Lorentzian \tab yes \tab yes \cr
5 \tab Pseudo-Voigt \tab yes \tab yes \cr
6 \tab Gaussian \tab no \tab no \cr
7 \tab Lorentzian \tab no \tab no \cr
8 \tab Pseudo-Voigt \tab no \tab no \cr
9 \tab Gaussian \tab no \tab yes \cr
10 \tab Lorentzian \tab no \tab yes \cr
11 \tab Pseudo-Voigt \tab no \tab yes
} The function for peakType = -1 is not an actual peak fit but rather the
Gaussian function (symmetric, no baseline) defined by the maximum intensity
point and its neighbors. It is used to automatically generate guess values
for position, width and amplitude if all peak parameters are set to 0.
}
\seealso{
Other Peak fitting functions: \code{\link{EvalMultiPeak}},
  \code{\link{EvalSinglePeak}},
  \code{\link{FitSinglePeak}}, \code{\link{MultiPeakFit}}
}
\concept{Peak fitting functions}
