% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetAcquisitionLogFromH5}
\alias{GetAcquisitionLogFromH5}
\title{Reads a single acquisition log entry.}
\usage{
GetAcquisitionLogFromH5(Filename, index)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{index}{Index of log entry.}
}
\value{
A list containing the timestamp (as a string) and log text.
}
\description{
\code{GetAcquisitionLogFromH5} reads a single acquisition log entry and
returns the timestamp and the log text.
}
\details{
The timestamp is the number of 100-nanosecond intervals since January 1,
1601 (UTC). Use \code{bit64::as.integer64(timestamp)} to convert the
timestamp string into a 64-bit integer value.
}
\examples{
\dontrun{
GetAcquisitionLogFromH5("path/to/file.h5", index = )
}
}

