% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetBufWriteProfile2FromH5}
\alias{GetBufWriteProfile2FromH5}
\title{Gets a linked buf/write profile.}
\usage{
GetBufWriteProfile2FromH5(Filename, PeakIndex, SegmentStartIndex,
  SegmentEndIndex)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{PeakIndex}{Index of peak to fetch buf/write profile from. All peaks are
read if \code{PeakIndex = -1}.}

\item{SegmentStartIndex}{Segment start index of data to fetch.}

\item{SegmentEndIndex}{Segment end index of data to fetch.}
}
\value{
A vector containing the buf/write profile(s).
}
\description{
\code{GetBufWriteProfile2FromH5} gets a linked buf/write profile for a given
peak (or all peaks) and segment slice.
}
\examples{
\dontrun{
GetBufWriteProfile2FromH5("path/to/file.h5", PeakIndex = -1,
SegmentStartIndex = 0, SegmentEndIndex = 0)
}
}

