% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetBufWriteProfileFromH5}
\alias{GetBufWriteProfileFromH5}
\title{Gets a linked buf/write profile.}
\usage{
GetBufWriteProfileFromH5(
  Filename,
  PeakIndex,
  SegmentStartIndex,
  SegmentEndIndex
)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{PeakIndex}{Index of peak to fetch buf/write profile from. All peaks are
read if \code{PeakIndex = -1}.}

\item{SegmentStartIndex}{Segment start index of data to fetch.}

\item{SegmentEndIndex}{Segment end index of data to fetch.}
}
\value{
A vector containing the buf/write profile(s).
}
\description{
\code{GetBufWriteProfileFromH5} gets a linked buf/write profile for a given
peak (or all peaks) and segment slice. If your data is not linked, use
\code{\link{GetPeakData}} to get the buf and/or write profiles.
}
\examples{
\dontrun{
GetBufWriteProfileFromH5("path/to/file.h5", PeakIndex = -1,
SegmentStartIndex = 0, SegmentEndIndex = 0)
}
}
