% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterBool}
\alias{GetDaqParameterBool}
\title{Gets a single boolean parameter.}
\usage{
GetDaqParameterBool(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}
}
\description{
\code{GetDaqParameterBool} gets a single boolean parameter.
}

