% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterBoolRef}
\alias{GetDaqParameterBoolRef}
\title{Gets a single boolean parameter.}
\usage{
GetDaqParameterBoolRef(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}
}
\description{
\code{GetDaqParameterBoolRef} gets a single boolean parameter.
}
\details{
This is the same as \code{GetDaqParameterBool}, but additionally it checks
for success and a TwRetVal string is returned if it is not sucessful.
}

