% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterDoubleRef}
\alias{GetDaqParameterDoubleRef}
\title{Gets a single double parameter.}
\usage{
GetDaqParameterDoubleRef(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}
}
\description{
\code{GetDaqParameterDoubleRef} gets a single double parameter.
}
\details{
This is the same as \code{GetDaqParameterDouble}, but additionally it checks
for success and a TwRetVal string is returned if it is not sucessful.
}
