% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterFloatRef}
\alias{GetDaqParameterFloatRef}
\title{Gets a single float parameter.}
\usage{
GetDaqParameterFloatRef(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\href{http://htmlpreview.github.io/?https://github.com/pasturm/TofDaqR/blob/master/tools/doc/TofDaqDll.htm#parameter_list}{TofDaq API documentation}
for a list of all available parameters.}
}
\description{
\code{GetDaqParameterFloatRef} gets a single float parameter.
}
\details{
This is the same as \code{GetDaqParameterFloat}, but additionally it checks
for success and a TwRetVal string is returned if it is not sucessful.
}
