% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterInt}
\alias{GetDaqParameterInt}
\title{Gets a single integer parameter.}
\usage{
GetDaqParameterInt(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}
}
\description{
\code{GetDaqParameterInt} gets a single integer parameter.
}
