% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterInt64}
\alias{GetDaqParameterInt64}
\title{Gets a single int64 parameter as a string.}
\usage{
GetDaqParameterInt64(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\href{http://htmlpreview.github.io/?https://github.com/pasturm/TofDaqR/blob/master/tools/doc/TofDaqDll.htm#parameter_list}{TofDaq API documentation}
for a list of all available parameters.}
}
\description{
\code{GetDaqParameterInt64} gets a single int64 parameter as a string.
}
\details{
The return string can be converted to integer64 using
\code{\link[bit64:as.integer64.character]{bit64::as.integer64()}}.
}
