% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterInt64Ref}
\alias{GetDaqParameterInt64Ref}
\title{Gets a single int64 parameter as a string.}
\usage{
GetDaqParameterInt64Ref(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}
}
\description{
\code{GetDaqParameterInt64Ref} gets a single int64 parameter as a string.
}
\details{
This is the same as \code{GetDaqParameterInt64}, but additionally it checks
for success and a TwRetVal string is returned if it is not sucessful.
}
