% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterIntRef}
\alias{GetDaqParameterIntRef}
\title{Gets a single integer parameter.}
\usage{
GetDaqParameterIntRef(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\href{https://htmlpreview.github.io/?https://github.com/pasturm/TofDaqR/blob/master/tools/doc/TofDaqDll.htm#parameter_list}{TofDaq API documentation}
for a list of all available parameters.}
}
\description{
\code{GetDaqParameterIntRef} gets a single integer parameter.
}
\details{
This is the same as \code{GetDaqParameterInt}, but additionally it checks
for success and a TwRetVal string is returned if it is not sucessful.
}
