% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterStringRef}
\alias{GetDaqParameterStringRef}
\title{Gets a single string parameter.}
\usage{
GetDaqParameterStringRef(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}
}
\description{
\code{GetDaqParameterStringRef} gets a single string parameter.
}
\details{
This is the same as \code{GetDaqParameter}, but returns \code{"TwInvalidValue"}
if the type of the parameter is not a string.
}

