% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDescriptor}
\alias{GetDescriptor}
\title{Gets various information about the active acquisition.}
\usage{
GetDescriptor()
}
\value{
A list containing the TSharedMemoryDesc structure
}
\description{
\code{GetDescriptor} retrieves the current TSharedMemoryDesc structure.
TSharedMemoryDesc contains various static information about the active
acquisition that can be retrieved by \code{GetDaqParameter} functions but
also information of DAQ progress.
See
\href{https://htmlpreview.github.io/?https://github.com/pasturm/TofDaqR/blob/master/tools/doc/TofDaqDll.htm}{TofDaq API documentation}
for more details.
}
\details{
int64 and unsigned int64 parameters are returned as string. They can be
converted to integer64 using \code{\link[bit64:as.integer64.character]{bit64::as.integer64()}}.
}
