% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwH5Dll.R
\name{GetDoubleAttributeFromH5}
\alias{GetDoubleAttributeFromH5}
\title{Reads a double attribute from the HDF5 file.}
\usage{
GetDoubleAttributeFromH5(filename, location, name)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}
}
\value{
A numeric attribute.
}
\description{
\code{GetDoubleAttributeFromH5} reads a double attribute from the HDF5 file.
}
\details{
Used to read attributes not available from \code{GetH5Descriptor}.
}
\examples{
\dontrun{
GetDoubleAttributeFromH5("path/to/file.h5", location = , name = )
}
}

