% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetEventListSpectrumFromH5}
\alias{GetEventListSpectrumFromH5}
\title{Reads the events of a spectrum from HDF5 data file.}
\usage{
GetEventListSpectrumFromH5(Filename, segmentIndex, bufIndex, writeIndex)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{segmentIndex}{Segment index.}

\item{bufIndex}{Buf index.}

\item{writeIndex}{Write index.}
}
\value{
A vector containing the event data.
}
\description{
\code{GetEventListSpectrumFromH5} reads the events of a single spectrum
given by segment, buf and write indices from HDF5 data file.
}
\details{
Note: only uncompressed event lists are supported.
}
\examples{
\dontrun{
GetEventListSpectrumFromH5("path/to/file.h5", segmentIndex = 0,
bufIndex = 0, writeIndex = 0)
}
}

