% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetFloatAttributeFromH5}
\alias{GetFloatAttributeFromH5}
\title{Reads a float attribute from the HDF5 file.}
\usage{
GetFloatAttributeFromH5(Filename, location, name)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}
}
\value{
A numeric attribute.
}
\description{
\code{GetFloatAttributeFromH5} reads a float attribute from the HDF5 file.
}
\details{
Used to read attributes not available from \code{GetH5Descriptor}.
}
\examples{
\dontrun{
GetFloatAttributeFromH5("path/to/file.h5", location = , name = )
}
}

