% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwH5Dll.R
\name{GetH5Descriptor}
\alias{GetH5Descriptor}
\title{Descriptor structure of Tofwerk HDF5 data file.}
\usage{
GetH5Descriptor(filename)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}
}
\value{
A list containing the \emph{TwH5Desc} structure.
}
\description{
\code{GetH5Descriptor} returns a descriptor structure for the Tofwerk HDF5 file.
}
\details{
The \emph{TwH5Desc} structure contains information about data dimensions,
available datasets and mass calibration. Additional attributes, which are not
available in the structure can be read using \code{Get...AttributeFromH5}
functions.
See \emph{/doc/TwH5Dll.htm} for more details.
}
\examples{
\dontrun{
GetH5Descriptor("path/to/file.h5")
}
}

