% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetInt64AttributeFromH5}
\alias{GetInt64AttributeFromH5}
\title{Reads a 64-bit integer attribute from the HDF5 file.}
\usage{
GetInt64AttributeFromH5(Filename, location, name)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}
}
\value{
An string attribute.
}
\description{
\code{GetInt64AttributeFromH5} reads a 64-bit integer attribute from the
HDF5 file.
}
\details{
Used to read attributes not available from \code{GetH5Descriptor}. int64
parameters are returned as string. They can be converted to integer64
using \code{\link[bit64]{as.integer64}}.
}
\examples{
\dontrun{
GetInt64AttributeFromH5("path/to/file.h5", location = , name = )
}
}

