% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetIntAttributeFromH5}
\alias{GetIntAttributeFromH5}
\title{Reads an integer attribute from the HDF5 file.}
\usage{
GetIntAttributeFromH5(Filename, location, name)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}
}
\value{
An integer attribute.
}
\description{
\code{GetIntAttributeFromH5} reads an integer attribute from the HDF5 file.
}
\details{
Used to read attributes not available from \code{GetH5Descriptor}.
}
\examples{
\dontrun{
GetIntAttributeFromH5("path/to/file.h5", location = , name = )
}
}

