% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwToolDll.R
\name{GetIsotopePattern}
\alias{GetIsotopePattern}
\title{Calculates the isotope pattern of a molecule.}
\usage{
GetIsotopePattern(molecule, abundanceLimit)
}
\arguments{
\item{molecule}{Molecule string.}

\item{abundanceLimit}{Absolute abundance limit for the generated pattern.}
}
\value{
List with mass and abundace vectors.
}
\description{
\code{GetIsotopePattern} parses the molecular formula and returns the
isotope pattern (mass and abundance).
}
\examples{
GetIsotopePattern("CO2", 1e-5)
}
\seealso{
Other Chemistry functions: \code{\link{GetIsotopePattern2}},
  \code{\link{GetMoleculeMass}}
}

