% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwToolDll.R
\name{GetIsotopePattern2}
\alias{GetIsotopePattern2}
\title{Calculates the isotope pattern of a molecule.}
\usage{
GetIsotopePattern2(molecule, abundanceLimit)
}
\arguments{
\item{molecule}{Molecule string.}

\item{abundanceLimit}{Absolute abundance limit for the generated pattern.}
}
\value{
List with mass and abundace vectors.
}
\description{
\code{GetIsotopePattern2} parses the molecular formula and returns the
isotope pattern (mass and abundance).
}
\details{
Same as \code{GetIsotopePattern} but using an exact algorithm and is
therefore suitable only for rather small molecules.
}
\examples{
GetIsotopePattern2("CO2", 1e-5)
}
\seealso{
Other Chemistry functions: \code{\link{GetIsotopePattern}},
  \code{\link{GetMoleculeMass}}
}

