% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetMassCalibInfo}
\alias{GetMassCalibInfo}
\title{Gets the description and number of parameters of the available mass
calibration functions.}
\usage{
GetMassCalibInfo(massCalibMode)
}
\arguments{
\item{massCalibMode}{Mass calibration mode (0 to 5).}
}
\value{
List with the description and number of calibration parameters for
the given \code{massCalibMode}.
}
\description{
\code{GetMassCalibInfo} gets the description and number of parameters of the
available mass calibration functions.
}
\details{
Note: Modes 3 and 4 are flawed. Don't use them. In mode 3 the fit does not
converge well, because of a bug (parameters not correctly initialized).
Mode 4 is two sequential fits, first mode 0, then a quadratic fit to the
residuals, which is an inferior implementation of mode 3. Mode 1 is for FTMS
data.
}
