% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwToolDll.R
\name{GetMoleculeMass}
\alias{GetMoleculeMass}
\title{Calculates the mass/charge ratio of a molecule.}
\usage{
GetMoleculeMass(molecule)
}
\arguments{
\item{molecule}{Molecule string.}
}
\value{
Mass/charge ratio.
}
\description{
\code{GetMoleculeMass} parses the molecular formula and returns the
mass/charge ratio of the molecule/atom/ion.
}
\examples{
GetMoleculeMass("CO2")
}
\seealso{
Other Chemistry functions: \code{\link{GetIsotopePattern2}},
  \code{\link{GetIsotopePattern}}
}

