% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwH5Dll.R
\name{GetPeakData}
\alias{GetPeakData}
\title{Gets data stored in /PeakData/PeakData from HDF5 data file.}
\usage{
GetPeakData(filename, peakOffset, peakCount, segOffset, segCount, bufOffset,
  bufCount, writeOffset, writeCount)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}

\item{peakOffset}{Peak offset.}

\item{peakCount}{Peak count.}

\item{segOffset}{Segment offset.}

\item{segCount}{Segment count.}

\item{bufOffset}{Buf offset.}

\item{bufCount}{Buf count.}

\item{writeOffset}{Write offset.}

\item{writeCount}{Write count.}
}
\description{
\code{GetPeakData} gets data stored in \code{/PeakData/PeakData} from HDF5 data file.
}

