% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetPeakData2}
\alias{GetPeakData2}
\title{Gets data stored in /PeakData2/PeakData from HDF5 data file.}
\usage{
GetPeakData2(
  Filename,
  peakOffset,
  peakCount,
  segOffset,
  segCount,
  bufOffset,
  bufCount,
  writeOffset,
  writeCount
)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{peakOffset}{Peak offset.}

\item{peakCount}{Peak count.}

\item{segOffset}{Segment offset.}

\item{segCount}{Segment count.}

\item{bufOffset}{Buf offset.}

\item{bufCount}{Buf count.}

\item{writeOffset}{Write offset.}

\item{writeCount}{Write count.}
}
\description{
\code{GetPeakData2} gets data stored in \code{/PeakData2/PeakData} from HDF5 data file.
}
