% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetPeakParameters}
\alias{GetPeakParameters}
\title{Gets parameters for a given peak.}
\usage{
GetPeakParameters(PeakIndex)
}
\arguments{
\item{PeakIndex}{Index of peak (zero-based numbering).}
}
\value{
A list with the peak paramters \emph{label}, \emph{mass}, \emph{loMass} and \emph{hiMass}.
}
\description{
\code{GetPeakParameters} gets parameters for a given peak.
}
