% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetPeakParametersFromH5}
\alias{GetPeakParametersFromH5}
\title{Peak parameters from HDF5 data file.}
\usage{
GetPeakParametersFromH5(Filename, PeakIndex = -1L)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{PeakIndex}{Index of peak. If index is -1 (default), peak parameters of all
peaks are read.}
}
\value{
A list with the peak paramters \emph{label}, \emph{mass}, \emph{loMass} and \emph{hiMass}.
}
\description{
\code{GetPeakParametersFromH5} reads peak parameters from the data file.
}
\examples{
\dontrun{
GetPeakParametersFromH5("path/to/file.h5")
}
}
