% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{GetRegUserDataDesc}
\alias{GetRegUserDataDesc}
\title{Reads the element descriptions of a registered data source.}
\usage{
GetRegUserDataDesc(Location)
}
\arguments{
\item{Location}{Location of group in HDF5 file identifying the registered data.}
}
\description{
\code{GetRegUserDataDesc} reads the element descriptions of a registered data source.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}

