% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRegUserDataSources}
\alias{GetRegUserDataSources}
\title{Queries names, dimensions and types of all registered data sources.}
\usage{
GetRegUserDataSources()
}
\value{
List with the location, nbrElements and type of the data sources.
type 1: data source values are written to disk for every write,
type 2: data source values are written to disk for every buf.
}
\description{
\code{GetRegUserDataSources} queries names, dimensions and types of all
data sources currently registered in TofDaq recorder.
}
\seealso{
Other Data storage functions: 
\code{\link{AddAttributeDouble}()},
\code{\link{AddAttributeInt}()},
\code{\link{AddAttributeString}()},
\code{\link{AddLogEntry}()},
\code{\link{AddUserDataMultiRow}()},
\code{\link{AddUserData}()},
\code{\link{GetRegUserDataDesc}()},
\code{\link{KeepFileOpen}()},
\code{\link{QueryRegUserDataSize}()},
\code{\link{ReadRegUserData}()},
\code{\link{RegisterUserDataBuf}()},
\code{\link{RegisterUserDataNoStore}()},
\code{\link{RegisterUserDataWrite}()},
\code{\link{UnregisterUserData}()},
\code{\link{UpdateUserData}()}
}
\concept{Data storage functions}
