% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRegUserDataSourcesFromH5}
\alias{GetRegUserDataSourcesFromH5}
\title{Lists all registered user datasets available in the data file.}
\usage{
GetRegUserDataSourcesFromH5(Filename)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}
}
\value{
A list containing the location, the length, whether is has a description
and the type of the data source dataset. type 1: data source values are
written to disk for every write, type 2: data source values are written to
disk for every buf.
}
\description{
\code{GetRegUserDataSourcesFromH5} lists all registered user data sets
available in the data file. Registered data sources can originate from data
source plugins, TofDaq recorder (e.g. DAQ temperatures) or data registered
through \code{RegisterUserData...} functions.
}
\examples{
\dontrun{
GetRegUserDataSourcesFromH5("path/to/file.h5")
}
}

