% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetSegmentProfile2FromH5}
\alias{GetSegmentProfile2FromH5}
\title{Segment profile from HDF5 data file.}
\usage{
GetSegmentProfile2FromH5(
  Filename,
  PeakIndex,
  BufStartIndex,
  BufEndIndex,
  WriteStartIndex,
  WriteEndIndex,
  BufWriteLinked = FALSE
)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{PeakIndex}{Index of peak to fetch segment profile from. All peaks are
read if \code{PeakIndex = -1}.}

\item{BufStartIndex}{Buf start index of data to fetch.}

\item{BufEndIndex}{Buf end index of data to fetch.}

\item{WriteStartIndex}{Write start index of data to fetch.}

\item{WriteEndIndex}{Write end index of data to fetch.}

\item{BufWriteLinked}{Indicating whether the buf and write dimension should
be considered linked or treated as independent dimensions (relevant only if
\code{WriteStartIndex != WriteEndIndex}). Default is \code{FALSE}.}
}
\value{
A vector containing the segment profile(s).
}
\description{
\code{GetSegmentProfile2FromH5} reads a segment profile for a given peak (or
all peaks) and buf and write slice.
}
\examples{
\dontrun{
GetSegmentProfile2FromH5("path/to/file.h5", PeakIndex = -1, BufStartIndex = 0,
BufEndIndex = 0, WriteStartIndex = 0, WriteEndIndex = 0)
}
}
