% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwH5Dll.R
\name{GetSegmentProfileFromH5}
\alias{GetSegmentProfileFromH5}
\title{Segment profile from HDF5 data file.}
\usage{
GetSegmentProfileFromH5(filename, PeakIndex, buf.start, buf.end, write.start,
  write.end, buf.write.linked = FALSE)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}

\item{PeakIndex}{Index of peak to fetch segment profile from. All peaks are
read if \code{PeakIndex = -1}.}

\item{buf.start}{Buf start index of data to fetch.}

\item{buf.end}{Buf end index of data to fetch.}

\item{write.start}{Write start index of data to fetch.}

\item{write.end}{Write end index of data to fetch.}

\item{buf.write.linked}{Indicating whether the buf and write dimension should
be considered linked or treated as independent dimensions (relevant only if
\code{write.start != write.end}). Default is \code{FALSE}.}
}
\value{
A vector containing the segment profile(s).
}
\description{
\code{GetSegmentProfileFromH5} reads a segment profile for a given peak (or
all peaks) and buf and write slice.
}
\examples{
\dontrun{
GetSegmentProfileFromH5("path/to/file.h5", PeakIndex = -1, buf.start = 0,
buf.end = 0, write.start = 0, write.end = 0)
}
}

