% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetSpecXaxisFromH5}
\alias{GetSpecXaxisFromH5}
\title{x-axis values of mass spectrum.}
\usage{
GetSpecXaxisFromH5(filename, type, writeIndex)
}
\arguments{
\item{writeIndex}{Write index to use for mass calibration (relevant only for
\code{abs(Type)== 1 or 2}). If the data file has no \emph{/TofData/MassCalibration}
dataset the standard mass calibration parameters are used (same for all
values of writeIndex).}

\item{Filename}{Path/filename of the HDF5 file.}

\item{Type}{x-axis type (0: sample index, 1: mass/charge [Th] (default),
-1: mass/charge [Th] (2nd TOF), 2: time of flight [microsec],
-2: time of flight [microsec] (2nd TOF), 3: frequency [kHz]).}
}
\value{
A vector containing the x-axis values.
}
\description{
\code{GetSpecXaxisFromH5} returns an array of x-axis values of the mass
spectrum.
}
\examples{
\dontrun{
GetSpecXaxisFromH5("path/to/file.h5")
}
}

