% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{GetSpecXaxisFromShMem}
\alias{GetSpecXaxisFromShMem}
\title{X-axis values of mass spectrum.}
\usage{
GetSpecXaxisFromShMem(Type)
}
\arguments{
\item{Type}{x-axis type (0: sample index, 1: mass/charge [Th],
-1: mass/charge [Th] (2nd TOF), 2: time of flight [microsec],
-2: time of flight [microsec] (2nd TOF), 3: frequency [kHz]).}
}
\value{
A vector containing the x-axis values.
}
\description{
\code{GetSpecXaxisFromShMem} returns an array of x-axis values of the mass
spectrum.
}

