% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetStickSpectrumFromH5}
\alias{GetStickSpectrumFromH5}
\title{Single (averaged) stick spectrum from HDF5 data file.}
\usage{
GetStickSpectrumFromH5(Filename, SegmentIndex, SegmentEndIndex, BufIndex,
  BufEndIndex, WriteIndex, WriteEndIndex, BufWriteLinked = FALSE,
  Normalize = TRUE)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{SegmentIndex}{Segment start index of data to fetch.}

\item{SegmentEndIndex}{Segment end index of data to fetch.}

\item{BufIndex}{Buf start index of data to fetch.}

\item{BufEndIndex}{Buf end index of data to fetch.}

\item{WriteIndex}{Write start index of data to fetch.}

\item{WriteEndIndex}{Write end index of data to fetch.}

\item{BufWriteLinked}{Indicating whether the buf and write dimension should
be considered linked or treated as independent dimensions (relevant only if
\code{WriteIndex != WriteEndIndex}). Default is \code{FALSE}.}

\item{Normalize}{If \code{FALSE} the spectrum is reported as sum,
if \code{TRUE} (default) the spectrum is normalized to counts per extraction.}
}
\value{
A vector containing the stick spectrum.
}
\description{
\code{GetStickSpectrumFromH5} reads a single stick spectrum (or an
averaged/summed hyperslab) from the HDF5 file.
}
\details{
If \code{SegmentIndex == SegmentEndIndex} and \code{BufIndex == BufEndIndex} and
\code{WriteIndex == WriteEndIndex} and \code{Normalize == FALSE} no
averaging/summing of spectra is done and the spectrum is reported as stored
in the dataset.
}
\examples{
\dontrun{
GetStickSpectrumFromH5("path/to/file.h5")
}
}

