% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{GetStickSpectrumFromShMem}
\alias{GetStickSpectrumFromShMem}
\title{Single stick spectrum from shared memory.}
\usage{
GetStickSpectrumFromShMem(SegmentIndex = 0, SegmentEndIndex = 0, BufIndex)
}
\arguments{
\item{SegmentIndex}{Segment start index of data to fetch.}

\item{SegmentEndIndex}{Segment end index of data to fetch.}

\item{BufIndex}{Buf index of data to fetch.}
}
\value{
A list containing the stick spectrum and corresponding masses.
}
\description{
\code{GetStickSpectrumFromShMem} reads a single stick spectrum from shared
memory.
}

