% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetSumSpectrumFromShMem}
\alias{GetSumSpectrumFromShMem}
\title{Sum spectrum from shared memory.}
\usage{
GetSumSpectrumFromShMem(Normalize = TRUE)
}
\arguments{
\item{Normalize}{If \code{FALSE} the spectrum is reported as sum,
if \code{TRUE} (default) the spectrum is normalized to counts per extraction.}
}
\description{
\code{GetSumSpectrumFromShMem} gets the sum spectrum from shared memory.
}
