% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetTofData}
\alias{GetTofData}
\title{Gets data stored in /FullSpectra/TofData from HDF5 data file.}
\usage{
GetTofData(Filename, sampleOffset, sampleCount, segOffset, segCount, bufOffset,
  bufCount, writeOffset, writeCount)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{sampleOffset}{Sample offset.}

\item{sampleCount}{Sample count.}

\item{segOffset}{Segment offset.}

\item{segCount}{Segment count.}

\item{bufOffset}{Buf offset.}

\item{bufCount}{Buf count.}

\item{writeOffset}{Write offset.}

\item{writeCount}{Write count.}
}
\description{
\code{GetTofData} gets data stored in \code{/FullSpectra/TofData} from HDF5 data file.
}

