% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetTofData2}
\alias{GetTofData2}
\title{Gets data stored in /FullSpectra2/TofData from HDF5 data file.}
\usage{
GetTofData2(
  Filename,
  sampleOffset,
  sampleCount,
  segOffset,
  segCount,
  bufOffset,
  bufCount,
  writeOffset,
  writeCount
)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{sampleOffset}{Sample offset.}

\item{sampleCount}{Sample count.}

\item{segOffset}{Segment offset.}

\item{segCount}{Segment count.}

\item{bufOffset}{Buf offset.}

\item{bufCount}{Buf count.}

\item{writeOffset}{Write offset.}

\item{writeCount}{Write count.}
}
\description{
\code{GetTofData2} gets data stored in \code{/FullSpectra2/TofData} from HDF5 data file.
}
