% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplementaryFunctions.R
\name{GetTofDataSinglePeak}
\alias{GetTofDataSinglePeak}
\title{Gets TofData of a single peak from a HDF5 data file.}
\usage{
GetTofDataSinglePeak(filename, peakIndex, massRange = NULL, secondTOF = FALSE)
}
\arguments{
\item{filename}{Path/filename.}

\item{peakIndex}{Peak index (one-based).}

\item{massRange}{Mass range in Th. Specified as a two-element vector
containing the low and high limit of the mass range. If \code{NULL} the mass
range is taken from the peak parameter list.}

\item{secondTOF}{If \code{TRUE} the data are read from /FullSpectra2/TofData}
}
\value{
A list containing the mass axis, time axis, TofData and averaged
spectrum of the peak, and the descriptor of the data file. TofData is a
matrix with the peak spectrum for every buf as columns and the unit of
TofData is mV (at the MCP signal out, before external preamps).
}
\description{
\code{GetTofDataSinglePeak} gets TofData and the averaged spectrum of a
single peak or a specified mass range from a HDF5 data file.
}
\details{
Note: this function is not part of the TofDaq API, but is included in the
package for convenience.
}
\seealso{
Other wrapper functions: 
\code{\link{FitTofDataSinglePeak}()}
}
\concept{wrapper functions}
