% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{GetTofSpectrumFromShMem}
\alias{GetTofSpectrumFromShMem}
\title{Single TOF spectrum from shared memory.}
\usage{
GetTofSpectrumFromShMem(SegmentIndex = 0, SegmentEndIndex = 0, BufIndex,
  Normalize = TRUE)
}
\arguments{
\item{SegmentIndex}{Segment start index of data to fetch (or -1 for complete
block copy).}

\item{SegmentEndIndex}{Segment end index of data to fetch (or -1 for complete
block copy).}

\item{BufIndex}{Buf index of data to fetch.}

\item{Normalize}{If \code{FALSE} the spectrum is reported as sum,
if \code{TRUE} (default) the spectrum is normalized to counts per extraction.}
}
\value{
A vector containing the mass spectrum or an array containing the
block of mass spectra if \code{SegmentIndex = SegmentEndIndex = -1}.
}
\description{
\code{GetTofSpectrumFromShMem} reads a single TOF spectrum (possibly
averaged/summed over segment dimension) from shared memory. If
\code{SegmentIndex = SegmentEndIndex = -1} the complete block of data is
copied and the \code{Normalize} flag is ignored.
}

