% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetUintAttributeFromH5}
\alias{GetUintAttributeFromH5}
\title{Reads an unsigned integer attribute from the HDF5 file.}
\usage{
GetUintAttributeFromH5(Filename, location, name)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}
}
\value{
An numeric attribute.
}
\description{
\code{GetUintAttributeFromH5} reads an unsigned integer attribute from the
HDF5 file.
}
\details{
Used to read attributes not available from \code{GetH5Descriptor}. Unsigned
integers are returned as numeric values in R.
}
\examples{
\dontrun{
GetUintAttributeFromH5("path/to/file.h5", location = , name = )
}
}
