% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetUserDataFromH5}
\alias{GetUserDataFromH5}
\title{Reads user data from the HDF5 file.}
\usage{
GetUserDataFromH5(Filename, location, rowIndex)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{rowIndex}{Row index.}
}
\value{
A list containing the row of user data and data description.
}
\description{
\code{GetUserDataFromH5} reads a row of user data added to the file using
the \code{AddUserData} function.
}
\details{
If you want to access data saved by a registered data source (or a data
 source plugin, which uses the same mechanism) use the \code{\link{GetRegUserDataFromH5}}
 function.
}
\examples{
\dontrun{
GetUserDataFromH5("path/to/file.h5", location = , rowIndex = )
}
}

