% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{H5AddLogEntry}
\alias{H5AddLogEntry}
\title{Adds an entry to an existing data file.}
\usage{
H5AddLogEntry(Filename, LogEntryText, LogEntryTime)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{LogEntryText}{Log text (max. 255 characters).}

\item{LogEntryTime}{Log entry time (number of 100-nanosecond intervals since
January 1, 1601 UTC, Windows FILETIME) passed as a string. Set it to "0" for
"now".}
}
\description{
\code{H5AddLogEntry} adds an entry to an existing data file. To add
acquisition log entries during a running acquisition use \code{AddLogEntry}.
}
