% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{H5SetMassCalib2Ex}
\alias{H5SetMassCalib2Ex}
\title{Changes the global mass calibration in the data file.}
\usage{
H5SetMassCalib2Ex(Filename, mode, nbrParams, p, mass, tof, weight, label)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{mode}{Mass calibration function to use.}

\item{nbrParams}{Number of mass calibration parameters.}

\item{p}{Vector with mass calibration parameters.}

\item{mass}{Vector with mass of the calibration points.}

\item{tof}{Vector with TOF sample index of the calibration points.}

\item{weight}{Vector with weight of the calibration points.}

\item{label}{Vector with labels/names/sum formula of the calibration points.}
}
\description{
\code{H5SetMassCalib2Ex} changes the global mass calibration in the data file.
If \code{nbrParams} is 0, the calibration parameters will be determined from
\code{mass}, \code{tof} and \code{weight}. If calibration parameters and
calibration points are provided, the calibration is given by the parameters
(no sanity check is performed whether the points yield the same parameters).
Labels to identify compound names/formulas used for calibration have a
maximum length of 255 characters.
}
\details{
\tabular{cl}{
mode \tab Mass calibration function \cr
0 \tab \eqn{i = p_1 \sqrt(m) + p_2} \cr
1 \tab \eqn{i = p_1/\sqrt(m) + p_2} \cr
2 \tab \eqn{i = p_1 m^{p_3} + p_2} \cr
3 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 (m - p_4)^2} \cr
4 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 m^2 + p_4 m + p_5} \cr
5 \tab \eqn{m = p_1 i^2 + p_2 i + p_3}
}
Note: Modes 3 and 4 are flawed. Don't use them. In mode 3 the fit does not
converge well, because of a bug (parameters not correctly initialized).
Mode 4 is two sequential fits, first mode 0, then a quadratic fit to the
residuals, which is an inferior implementation of mode 3. Mode 1 is for FTMS
data.
}
