% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{H5SetMassCalibDynamic}
\alias{H5SetMassCalibDynamic}
\title{Stores dynamic mass calibration for a given spectrum in the data file.}
\usage{
H5SetMassCalibDynamic(filename, writeIndex, par, stat)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}

\item{writeIndex}{Write index of the calibration to store.}

\item{par}{Numeric vector holding the actual calibration values.}

\item{stat}{Numeric vector with additional information on how the parameters
were obtained (typically).}
}
\description{
\code{H5SetMassCalibDynamic} stores dynamic mass calibration for a given
spectrum in the data file.
}
\details{
This function can be used to delete the dynamic calibration information by
passing the special parameter set: \code{writeIndex} = -1, \code{nbrParams}
= \code{nbrStat} = 0 and \code{par} = \code{stat} = \code{NULL}. With the
\code{stat} array additional information can be stored in the dataset
\code{/FullSpectra/MassCalibrationStats} (no official format definition or
supporting API functions).
}
