% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InitializeDll}
\alias{InitializeDll}
\title{Initializes the TofDaqDll.dll.}
\usage{
InitializeDll()
}
\description{
\code{InitializeDll} initializes the TofDaqDll.dll. It is usually not necessary
to call \code{InitializeDll} explicitely, as it is called automatically by
functions that need the DLL to be in an initialized state.
}

