% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IssueDio4Pulse}
\alias{IssueDio4Pulse}
\title{Issues a TTL pulse on the digital output line 4.}
\usage{
IssueDio4Pulse(delay, width)
}
\arguments{
\item{delay}{Delay before issuing pulse in ms.}

\item{width}{Pulse width in ms.}
}
\description{
\code{IssueDio4Pulse} issues a TTL pulse on the digital output line 4
specified by a delay and a pulse width.
}
\details{
Note that in order for this command to work the Dio4Mode parameter must be
set to 2 (pulsed) or 3 (manual).
}
