% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{KeepFileOpen}
\alias{KeepFileOpen}
\title{Allows to keep the data file open at the end of an acquisition.}
\usage{
KeepFileOpen(keepOpen)
}
\arguments{
\item{keepOpen}{Issue \code{TRUE} (after DAQ start) to signal to recorder to
keep the data file open. When done adding data, issue \code{FALSE} to allow
the recorder to close the file.}
}
\description{
\code{KeepFileOpen} allows to keep the data file open at the end of an acquisition.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddLogEntry}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}

