% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwToolDll.R
\name{Mass2Tof}
\alias{Mass2Tof}
\title{Converts from mass/charge to sample index.}
\usage{
Mass2Tof(mass, massCalibMode = 0, p)
}
\arguments{
\item{mass}{Vector of mass/charge values to convert.}

\item{massCalibMode}{Mass calibration function to use. See below.}

\item{p}{Vector containing the calibration parameters (number depends on
\code{MassCalibMode}, see below).}
}
\value{
Sample indices.
}
\description{
\code{Mass2Tof} converts from mass/charge to sample index.
}
\details{
\tabular{cl}{
massCalibMode \tab Mass calibration function \cr
0 \tab \eqn{i = p_1 \sqrt(m) + p_2} \cr
1 \tab \eqn{i = p_1/\sqrt(m) + p_2} \cr
2 \tab \eqn{i = p_1 m^{p_3} + p_2} \cr
3 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 (m - p_4)^2} \cr
4 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 m^2 + p_4 m + p_5} \cr
5 \tab \eqn{m = p_1 i^2 + p_2 i + p_3}
}
}
\examples{
Mass2Tof(100, massCalibMode = 0, p = c(3,5))
}
\seealso{
\code{\link{Tof2Mass}}
}

