% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MassCalibrate}
\alias{MassCalibrate}
\title{Performs a mass calibration.}
\usage{
MassCalibrate(massCalibMode, mass, tof, weight = NULL)
}
\arguments{
\item{massCalibMode}{Mass calibration function to use. See below.}

\item{mass}{Vector of mass/charge values.}

\item{tof}{Vector of sample indices (of same length as \code{mass}).}

\item{weight}{Vector of weights (if \code{NULL} (default) all weights are set equal).}
}
\value{
Vector of calibration parameters.
}
\description{
\code{MassCalibrate} performs a mass calibration for a list of
mass/sample index/weight values and for a given calibration mode.
}
\details{
\tabular{cl}{
massCalibMode \tab Mass calibration function \cr
0 \tab \eqn{i = p_1 \sqrt(m) + p_2} \cr
1 \tab \eqn{i = p_1/\sqrt(m) + p_2} \cr
2 \tab \eqn{i = p_1 m^{p_3} + p_2} \cr
3 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 (m - p_4)^2} \cr
4 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 m^2 + p_4 m + p_5} \cr
5 \tab \eqn{m = p_1 i^2 + p_2 i + p_3}
}
Note: Modes 3 and 4 are flawed. Don't use them. In mode 3 the fit does not
converge well, because of a bug (parameters not correctly initialized).
Mode 4 is two sequential fits, first mode 0, then a quadratic fit to the
residuals, which is an inferior implementation of mode 3. Mode 1 is for FTMS
data.
}

