% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MatchSpectra}
\alias{MatchSpectra}
\title{Checks two spectra for similarity.}
\usage{
MatchSpectra(spec1, spec2, matchMethod = 0L)
}
\arguments{
\item{spec1}{First spectrum to match.}

\item{spec2}{Second spectrum to match.}

\item{matchMethod}{Method to use. Currently only method 0 is implemented.}
}
\value{
Match score in units of percent.
}
\description{
\code{MatchSpectra} matches two spectra and returns a similarity score
(0 - 100 %).
}
\details{
Values < 0.0 in spec1 or spec2 will be set to 0.0. Currently, only
matchMethod 0 is implemented and relies on the euclidean distance between
the spectra.
}
