% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MultiPeakFit}
\alias{MultiPeakFit}
\title{Performs a multi-peak fit.}
\usage{
MultiPeakFit(dataX, dataY, mass, intensity, commonPar, options)
}
\arguments{
\item{dataX}{x axis data.}

\item{dataY}{y axis data.}

\item{mass}{Vector of the known positions of the peaks.}

\item{intensity}{Vector of initial guesses for intensities. If all values
are 0 the guess values for peak intensities are generated automatically.}

\item{commonPar}{Vector of guess values of common peak parameters. \code{commonPar}
has 6 elements: [1] offset of common baseline, [2] slope of common baseline,
[3] left FWHM, [4] right FWHM (same as left FWHM for symmetric peaks),
[5] shape parameter mu (applies only for peak model Pseudo-Voigt),
[6] common mass shift of peaks. \code{options} determines which of the
common parameters are optimized.}

\item{options}{List of peak model and optimization options (see Details).}
}
\value{
List with the optimized intensities and common peak parameters.
}
\description{
\code{MultiPeakFit} performs a multi-peak fit for (partially) overlapping
peaks. All peaks in a multi-peak cluster share common peak parameters except
amplitude and position. Various options allow for a more or less constrained
fit (see description of options below).
}
\details{
The peak positions are not optimized, but the common mass shift parameter in
\code{commonPar} allows for minimal refinement of the peak positions due to
imperfect mass calibration.

\code{options} is a list containing:
\tabular{rcl}{
peakModel \tab = \tab 0 (Gauss) or 1 (Lorentz) or 2 (Pseudo-Voigt) \cr
asymmetric \tab = \tab 0 (symmetric peak model) or 1 (asymmetric peak model) \cr
baseline \tab = \tab 0 (no optimization of baseline parameters) or 1 (optimize baseline parameters) \cr
width \tab = \tab 0 (no optimization of width parameters) or 1 (optimize width parameters) \cr
peakShape \tab = \tab 0 (no optimization of peak shape (mu) parameter) or 1 (optimize peak shape (mu) parameter (applies only to pseudo-Voigt)) \cr
massShift \tab = \tab 0 (no optimization of common mass shift parameters) or 1 (optimize common mass shift parameters) \cr
amplitude \tab = \tab 0 (no constraint on amplitudes) or 1 (constrain sum of baseline and all peaks to total intensity in spectrum)
} Note that if a given parameter is not activated for optimization, the
supplied (guess) values are still used (e.g. specify a known baseline
without optimizing the parameters).
}
\seealso{
Other Peak fitting functions: \code{\link{EvalMultiPeak}},
  \code{\link{EvalSinglePeak}},
  \code{\link{FitSinglePeak2}}, \code{\link{FitSinglePeak}}
}
