% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{OnDemandMassCalibration}
\alias{OnDemandMassCalibration}
\title{Arms/executes on demand mass calibration.}
\usage{
OnDemandMassCalibration(action)
}
\arguments{
\item{action}{0: arms mass calibration, 1: updates mass calibration using
data acquired since previous arm.}
}
\description{
\code{OnDemandMassCalibration} arms/executes on demand mass calibration.
Requires that parameter \code{ReCalibFreq} is set to 3 (on demand) in order to work.
}

