% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{QueryRegUserDataSize}
\alias{QueryRegUserDataSize}
\title{Queries the size of a registered data source.}
\usage{
QueryRegUserDataSize(Location)
}
\arguments{
\item{Location}{Location of group in HDF5 file identifying the registered data.}
}
\description{
\code{QueryRegUserDataSize} queries the size (nbrElements) of a registered data source.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddLogEntry}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}

