% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadRegUserData}
\alias{ReadRegUserData}
\title{Reads the current values of a registered data source.}
\usage{
ReadRegUserData(Location, NbrElements)
}
\arguments{
\item{Location}{Location of group in HDF5 file where the datasets are created.}

\item{NbrElements}{Number of elements to read.}
}
\value{
Vector containing the registered user data.
}
\description{
\code{ReadRegUserData} reads the current values of a registered data source.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddLogEntry}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}
