% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RegisterUserDataBuf}
\alias{RegisterUserDataBuf}
\title{Registers a data source to store (synchronous) user supplied data.}
\usage{
RegisterUserDataBuf(Location, NbrElements, ElementDescription = NULL,
  CompressionLevel = 0L)
}
\arguments{
\item{Location}{Location of group in HDF5 file where the datasets are created.}

\item{NbrElements}{Number of elements to store per buf.}

\item{ElementDescription}{Vector of length \code{NbrElements} containing the
text description of elements. If \code{ElementDescription} is \code{NULL}
the dataset "TwInfo" is not created.}

\item{CompressionLevel}{Compression level used for data storage (0: no
compression, 1-9: increasing levels of compression (and CPU load)).}
}
\description{
\code{RegisterUserDataBuf} registers a data source to store user supplied
data synchronously to the TOF data acquistion (every buf) into the data file
being currently recorded. Creates datasets "TwData" and "TwInfo" at \code{Location}.
}
\details{
Needs to be executed before starting the acquisition.
Use \code{\link{UpdateUserData}} to actually store the data.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddLogEntry}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}
