% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RegisterUserDataNoStore}
\alias{RegisterUserDataNoStore}
\title{Registers a data source for (synchronous) user supplied data.}
\usage{
RegisterUserDataNoStore(Location, NbrElements, ElementDescription = NULL)
}
\arguments{
\item{Location}{Location of group in HDF5 file where the datasets are created.}

\item{NbrElements}{Number of elements to store per write.}

\item{ElementDescription}{Vector of length \code{NbrElements} containing the
text description of elements. If \code{ElementDescription} is \code{NULL}
the dataset "TwInfo" is not created.}
}
\description{
\code{RegisterUserDataNoStore} registers a data for user supplied
data (synchronous to the TOF data acquistion) but the data is not stored
in the data file.
}
\details{
Needs to be executed before starting the acquisition.
Use \code{\link{UpdateUserData}} to update the data.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddLogEntry}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}
