% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{ReleaseSharedMemory}
\alias{ReleaseSharedMemory}
\title{Manually releases the shared memory acquisition buffers.}
\usage{
ReleaseSharedMemory()
}
\description{
\code{ReleaseSharedMemory} manually releases the shared memory acquisition
buffers. This is needed if \code{\link{KeepSharedMemMapped}} has been set to
\code{TRUE}.
}

