% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SaturationWarning}
\alias{SaturationWarning}
\title{Checks if the signal is saturating the DAQ system.}
\usage{
SaturationWarning()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{SaturationWarning} checks if the signal is saturating the analog input
of the DAQ system.
}
\details{
Signals saturation of the input signal (saturation due to analog input
limitations, not MCP limit or non-linearity in signal which may happen at
significantly lower signal values).
}
