% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SaveIniFile}
\alias{SaveIniFile}
\title{Saves the current configuration (*.ini) to disk.}
\usage{
SaveIniFile(IniFile = NULL)
}
\arguments{
\item{IniFile}{Path/filename of the configuration file. If no path is
specified, the file will be saved in the TofDaq recorder directory.
If \code{IniFile} is an empty string or \code{NULL}, "TwApiTmpIni.ini"
will be used. If a path is specified, existing files cannot be overwritten.}
}
\description{
\code{SaveIniFile} saves the current configuration (*.ini) to disk.
}
