% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplementaryFunctions.R
\name{SaveMassTableToFile}
\alias{SaveMassTableToFile}
\title{Saves the current peak parameters to a file.}
\usage{
SaveMassTableToFile(filename = "TmpMassTable.txt")
}
\arguments{
\item{filename}{Path/filename. If not specified, "TmpMassTable.txt" in the
current working directory will be used.}
}
\description{
\code{SaveMassTableToFile} saves the current peak parameters to a file.
}
\details{
Note: this function is not part of the TofDaq API, but is included in the
package for convenience.
}
