% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SendDioStartSignal}
\alias{SendDioStartSignal}
\title{Sends a digital start signal.}
\usage{
SendDioStartSignal()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{SendDioStartSignal} sends a digital start signal.
}
\details{
Software override of digital start signal as configured with DioStart...
TofDaq recorder parameters. See also \code{\link{WaitingForDioStartSignal}}
for finding out when this function can be called successfully.
}
