% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetDaqParameterBool}
\alias{SetDaqParameterBool}
\title{Sets a single parameter with a boolean value.}
\usage{
SetDaqParameterBool(Parameter, Value)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}

\item{Value}{\code{TRUE} or \code{FALSE}.}
}
\description{
\code{SetDaqParameterBool} sets a single parameter with a boolean value.
}
