% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetDaqParameterDouble}
\alias{SetDaqParameterDouble}
\title{Sets a single parameter with a double value.}
\usage{
SetDaqParameterDouble(Parameter, Value)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}

\item{Value}{Numeric value.}
}
\description{
\code{SetDaqParameterDouble} sets a single parameter with a double value.
}

