% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetDio4State}
\alias{SetDio4State}
\title{Switches the digital output line 4 between states.}
\usage{
SetDio4State(state)
}
\arguments{
\item{state}{0: idle state, 1 (or any value other than 0) active state.}
}
\description{
\code{SetDio4State} switches the digital output line 4 between states.
}
\details{
Note that in order for this command to work the Dio4Mode parameter must be
set to 2 (pulsed) or 3 (manual).
}
