% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetDoubleAttributeInH5}
\alias{SetDoubleAttributeInH5}
\title{Writes a double attribute to the HDF5 file.}
\usage{
SetDoubleAttributeInH5(Filename, location, name, attribute)
}
\arguments{
\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}

\item{attribute}{Double attribute.}

\item{filename}{Path/filename of the HDF5 file.}
}
\description{
\code{SetDoubleAttributeInH5} writes a double attribute to the HDF5 file.
}
\examples{
\dontrun{
SetDoubleAttributeInH5("path/to/file.h5", location = , name = , attribute = )
}
}

