% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetInt64AttributeInH5}
\alias{SetInt64AttributeInH5}
\title{Writes an int64 attribute to the HDF5 file.}
\usage{
SetInt64AttributeInH5(Filename, location, name, attribute)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}

\item{attribute}{Int64 attribute passed as a string.}
}
\description{
\code{SetInt64AttributeInH5} writes an int64 attribute to the HDF5 file.
}
\examples{
\dontrun{
SetInt64AttributeInH5("path/to/file.h5", location = , name = , attribute = )
}
}

