% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwH5Dll.R
\name{SetIntAttributeInH5}
\alias{SetIntAttributeInH5}
\title{Writes an integer attribute to the HDF5 file.}
\usage{
SetIntAttributeInH5(filename, location, name, attribute)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}

\item{attribute}{Integer attribute.}
}
\description{
\code{SetIntAttributeInH5} writes an integer attribute to the HDF5 file.
}
\examples{
\dontrun{
SetIntAttributeInH5("path/to/file.h5", location = , name = , attribute = )
}
}

