% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetMassCalib2Ex}
\alias{SetMassCalib2Ex}
\title{Configures the mass calibration that will be used for the next acquisition.}
\usage{
SetMassCalib2Ex(mode, nbrParams, p, mass, tof, weight, label)
}
\arguments{
\item{mode}{Mass calibration function to use.}

\item{nbrParams}{Number of mass calibration parameters.}

\item{p}{Vector with mass calibration parameters.}

\item{mass}{Vector with mass of the calibration points.}

\item{tof}{Vector with TOF sample index of the calibration points.}

\item{weight}{Vector with weight of the calibration points.}

\item{label}{Vector with labels of the calibration points.}
}
\description{
\code{SetMassCalib2Ex} configures the mass calibration that will be used for
the next acquisition(s). If \code{nbrParams} is 0, the calibration parameters are
determined by the TofDaq recorder based on the mass, tof and weight arrays.
If calibration parameters and calibration point information is supplied the
calibration parameters define the calibration (no "sanity" check is
performed whether the point information yields the same mass calibration
parameters). Labels to identify compound names/formulas used for
calibration have a maximum length of 255 characters.
}
\details{
\tabular{cl}{
mode \tab Mass calibration function \cr
0 \tab \eqn{i = p_1 \sqrt(m) + p_2} \cr
1 \tab \eqn{i = p_1/\sqrt(m) + p_2} \cr
2 \tab \eqn{i = p_1 m^{p_3} + p_2} \cr
3 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 (m - p_4)^2} \cr
4 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 m^2 + p_4 m + p_5} \cr
5 \tab \eqn{m = p_1 i^2 + p_2 i + p_3}
}
Note: Modes 3 and 4 are flawed. Don't use them. In mode 3 the fit does not
converge well, because of a bug (parameters not correctly initialized).
Mode 4 is two sequential fits, first mode 0, then a quadratic fit to the
residuals, which is an inferior implementation of mode 3. Mode 1 is for FTMS
data.
}
