% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetStringAttributeInH5}
\alias{SetStringAttributeInH5}
\title{Writes a string attribute to the HDF5 file.}
\usage{
SetStringAttributeInH5(Filename, location, name, attribute)
}
\arguments{
\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}

\item{attribute}{String attribute (max. 256 characters).}

\item{filename}{Path/filename of the HDF5 file.}
}
\description{
\code{SetStringAttributeInH5} writes a string attribute to the HDF5 file.
}
\examples{
\dontrun{
SetIntAttributeInH5("path/to/file.h5", location = , name = , attribute = )
}
}

