% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetTimeout}
\alias{SetTimeout}
\title{Sets the timeout.}
\usage{
SetTimeout(timeout)
}
\arguments{
\item{timeout}{Timeout in ms. Default is 500 ms.}
}
\description{
\code{SetTimeout} sets the global timeout for all functions that can time
out. Default is 500 ms.
}

