% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetUintAttributeInH5}
\alias{SetUintAttributeInH5}
\title{Writes an unsigned integer attribute to the HDF5 file.}
\usage{
SetUintAttributeInH5(Filename, location, name, attribute)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}

\item{attribute}{Unsigned integer attribute (passed as numeric value).}
}
\description{
\code{SetUintAttributeInH5} writes an unsigned integer attribute to the HDF5 file.
}
\examples{
\dontrun{
SetUintAttributeInH5("path/to/file.h5", location = , name = , attribute = )
}
}
