% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{ShowConfigWindow}
\alias{ShowConfigWindow}
\title{Shows the TofDaq recorder configuration windows.}
\usage{
ShowConfigWindow(ConfigWindowIndex = 0)
}
\arguments{
\item{ConfigWindowIndex}{Index of configuration tab to show (valid range: 0-6)}
}
\description{
\code{ShowConfigWindow} shows the different tabs of the TofDaq recorder
configuration window.
}

