% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SiCleanup}
\alias{SiCleanup}
\title{Cleans up the state in the DLL.}
\usage{
SiCleanup()
}
\description{
\code{SiCleanup} cleans up the state in the DLL. After a call to this
function \code{\link{SiInitializeHistograms}} can be called again.
}
