% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SiEvalPhd}
\alias{SiEvalPhd}
\title{Evaluates the fitted single ion distribution.}
\usage{
SiEvalPhd(par, intensity)
}
\arguments{
\item{par}{Vector of fitted parameter values.}

\item{intensity}{Vector of intensities (histogram x-axis data).}
}
\value{
Vector with y-axis data.
}
\description{
\code{SiEvalPhd} evaluates the fitted single ion distribution.
}
\seealso{
Other Single ion histogramming functions: \code{\link{SiCleanup}},
  \code{\link{SiFitPhd}}, \code{\link{SiFitRateFromPhd}},
  \code{\link{SiGetHistogram}},
  \code{\link{SiGetSumHistogram}},
  \code{\link{SiInitializeHistograms}},
  \code{\link{SiProcessSpectrumFromShMem}},
  \code{\link{SiProcessSpectrum}},
  \code{\link{SiResetHistograms}},
  \code{\link{SiSetProcessingOptions}}
}
