% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwTool.R
\name{SiFitPhd}
\alias{SiFitPhd}
\title{Fits a (slightly modified) log-normal distribution to the histogram.}
\usage{
SiFitPhd(intensity, counts)
}
\arguments{
\item{intensity}{Vector of intensities (histogram x-axis data).}

\item{counts}{Vector of counts (histogram y-axis data).}
}
\value{
A list with the FWHM of the distribution, the position of the fitted
maximum and a vector with the values of the four fitting parameters.
}
\description{
\code{SiFitPhd} fits a (slightly modified) log-normal distribution to the
histogram data.
}
\details{
The following equation with the parameters A, w, x0 and xc is fitted: \cr
\eqn{A/(\sqrt(2*pi)*w*(x-x0))*exp(-(log((x-x0)/xc)^2)/(2*w^2))}
}
\seealso{
Other Single ion histogramming functions.: \code{\link{SiCleanup}},
  \code{\link{SiEvalPhd}}, \code{\link{SiFitRateFromPhd}},
  \code{\link{SiGetHistogram}},
  \code{\link{SiGetSumHistogram}},
  \code{\link{SiInitializeHistograms}},
  \code{\link{SiProcessSpectrumFromShMem}},
  \code{\link{SiProcessSpectrum}},
  \code{\link{SiResetHistograms}},
  \code{\link{SiSetProcessingOptions}}
}

