% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SiFitRateFromPhd}
\alias{SiFitRateFromPhd}
\title{Fits an event rate to a multi-ion histogram.}
\usage{
SiFitRateFromPhd(intensity, counts, siPar)
}
\arguments{
\item{intensity}{Vector of intensities (histogram x-axis data).}

\item{counts}{Vector of counts (histogram y-axis data).}

\item{siPar}{Vector with fitted parameter values.}
}
\value{
A list with the fitted rate (ions/extraction) and vector with the
fitted multi-ion distribution.
}
\description{
\code{SiFitRateFromPhd} takes a fitted single ion distribution as input
and fits an event rate to a multi-ion histogram (intensity and counts)
assuming poisson distribution of n-ion events.
}
\seealso{
Other Single ion histogramming functions: \code{\link{SiCleanup}},
  \code{\link{SiEvalPhd}}, \code{\link{SiFitPhd}},
  \code{\link{SiGetHistogram}},
  \code{\link{SiGetSumHistogram}},
  \code{\link{SiInitializeHistograms}},
  \code{\link{SiProcessSpectrumFromShMem}},
  \code{\link{SiProcessSpectrum}},
  \code{\link{SiResetHistograms}},
  \code{\link{SiSetProcessingOptions}}
}
