% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwToolDll.R
\name{SiGetHistogram}
\alias{SiGetHistogram}
\title{Gets a histogram of the single ion intensities.}
\usage{
SiGetHistogram(histogramIndex)
}
\arguments{
\item{histogramIndex}{Index (0-based) of the histogram. It corresponds to the
mass range defined with \code{\link{SiInitializeHistograms}}.}
}
\value{
A list with the intensities (histogram x-values), counts (histogram
y-values), the number of spectra that were processed for this histogram and
the mean histogram value i.e. sum(intensity[i]*counts[i])/sum(counts[i]).
}
\description{
\code{SiGetHistogram} gets a histogram of the single ion intensities for a
mass range defined with \code{\link{SiInitializeHistograms}}.
}
\details{
Note: R crashes if \code{histogramIndex} is set to max(histogramIndex)+1
(API bug).
}
\seealso{
Other Single ion histogramming functions.: \code{\link{SiCleanup}},
  \code{\link{SiEvalPhd}}, \code{\link{SiFitPhd}},
  \code{\link{SiFitRateFromPhd}},
  \code{\link{SiGetSumHistogram}},
  \code{\link{SiInitializeHistograms}},
  \code{\link{SiProcessSpectrumFromShMem}},
  \code{\link{SiProcessSpectrum}},
  \code{\link{SiResetHistograms}},
  \code{\link{SiSetProcessingOptions}}
}

