% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SiGetSumHistogram}
\alias{SiGetSumHistogram}
\title{Gets a sum histogram of the single ion intensities.}
\usage{
SiGetSumHistogram(specType, minMass, maxMass, minRate, maxRate)
}
\arguments{
\item{specType}{Spectrum type index (non-negative integer).}

\item{minMass}{Minimum mass for histogram filtering.}

\item{maxMass}{Maximum mass for histogram filtering.}

\item{minRate}{Minimum event count rate for histogram filtering.}

\item{maxRate}{Maximum event count rate for histogram filtering.}
}
\value{
A list with the intensities (histogram x-values), counts (histogram
y-values), the number of spectra that were processed for this histogram and
the mean histogram value i.e. sum(intensity[i]*counts[i])/sum(counts[i]).
}
\description{
\code{SiGetSumHistogram} gets a histogram of the single ion intensities, which
is a sum over all histograms of a given \code{specType} within the rate and
mass range as specified by \code{minMass}, \code{maxMass}, \code{minRate} and \code{maxRate}.
}
\seealso{
Other Single ion histogramming functions.: \code{\link{SiCleanup}},
  \code{\link{SiEvalPhd}}, \code{\link{SiFitPhd}},
  \code{\link{SiFitRateFromPhd}},
  \code{\link{SiGetHistogram}},
  \code{\link{SiInitializeHistograms}},
  \code{\link{SiProcessSpectrumFromShMem}},
  \code{\link{SiProcessSpectrum}},
  \code{\link{SiResetHistograms}},
  \code{\link{SiSetProcessingOptions}}
}

