% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwTool.R
\name{SiInitializeHistograms}
\alias{SiInitializeHistograms}
\title{Initializes the single ion histogramming.}
\usage{
SiInitializeHistograms(loMass, hiMass, specType = NULL)
}
\arguments{
\item{loMass}{Vector of the lower borders of the mass ranges.}

\item{hiMass}{Vector of the upper borders of the mass ranges.}

\item{specType}{Vector of spectrum type indices (non-negative integers). If
specType is \code{NULL}, all mass ranges get a default spectrum type of 0.}
}
\description{
\code{SiInitializeHistograms} initializes the single ion histogramming.
}
\details{
This function must be called before calling the functions
\code{\link{SiSetProcessingOptions}}, \code{\link{SiProcessSpectrum}}, \code{\link{SiGetHistogram}},
\code{\link{SiResetHistograms}} or \code{\link{SiCleanup}}. Each mass range specified by
\code{loMass} and \code{hiMass} elements is associated with a spectrum type that allows to
get separate statistics for multi-spectrum acquisitions (bipolar or pulsed
experiments).
}
\seealso{
Other Single ion histogramming functions.: \code{\link{SiCleanup}},
  \code{\link{SiEvalPhd}}, \code{\link{SiFitPhd}},
  \code{\link{SiFitRateFromPhd}},
  \code{\link{SiGetHistogram}},
  \code{\link{SiGetSumHistogram}},
  \code{\link{SiProcessSpectrumFromShMem}},
  \code{\link{SiProcessSpectrum}},
  \code{\link{SiResetHistograms}},
  \code{\link{SiSetProcessingOptions}}
}

