% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SiProcessSpectrum}
\alias{SiProcessSpectrum}
\title{Processes a spectrum.}
\usage{
SiProcessSpectrum(spectrum, specType)
}
\arguments{
\item{spectrum}{Vector holding the spectrum to process.}

\item{specType}{Spectrum type index (non-negative integer).}
}
\value{
A list with the baseline and threshold value.
}
\description{
\code{SiProcessSpectrum} processes a spectrum according to the options set for
it's spectrum type.
}
