% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SiResetHistograms}
\alias{SiResetHistograms}
\title{Resets all histograms and spectrum counters to zero.}
\usage{
SiResetHistograms()
}
\description{
\code{SiResetHistograms} resets all histograms and spectrum counters to zero.
}
