% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tof}
\alias{tof}
\title{Time-of-flight calculation.}
\usage{
tof(toftype = "LTOF", drift = 6000, pulse = 1000, mass = 100,
  x = as.numeric(c(0)), v = as.numeric(c(0)))
}
\arguments{
\item{toftype}{TOF type (\code{"CTOF"}, \code{"HTOF"}, \code{"HTOF-W"} or \code{"LTOF"})}

\item{drift}{Drift tube voltage (V).}

\item{pulse}{Extraction pulse voltage (V).}

\item{mass}{Mass-to-charge ratio (Th).}

\item{x}{Initial position deviation of the ion(s) from the extraction plane (m).}

\item{v}{Initial velocity of the ion(s) in extraction direction (m/s).}
}
\value{
Time-of-flight (s)
}
\description{
\code{tof} calculates the time-of-flight of ions in an ideal two-stage
extraction/two-stage reflection TOFMS (with second-order space focusing).
}
\details{
\code{"HTOF-W"} is the W-mode and assumes hard mirror voltage = 1.5*pulse voltage.
\code{x} and \code{v} can be vectors, but must be the same length.

Note: this function is not part of the TofDaq API, but is included in the
package for convenience.
}
\references{
Laiko, V.V. and A.F. Dodonov (1994), Resolution and Spectral-line
Shapes in the Reflecting Time-of-flight Mass Spectrometer with Orthogonally
Injected Ions, \emph{Rapid Communications in Mass Spectrometry}, \strong{8},
720-726.
}
\keyword{internal}

