% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TofDaqRunning}
\alias{TofDaqRunning}
\title{Checks if TofDaq recorder application is running.}
\usage{
TofDaqRunning()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{TofDaqRunning} checks if TofDaq recorder application is running.
}
