% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupplementaryFunctions.R
\docType{data}
\name{Tps1rc}
\alias{Tps1rc}
\title{TPS1 RC codes and names.}
\format{
An object of class \code{data.frame} with 78 rows and 2 columns.
}
\usage{
Tps1rc
}
\description{
Data frame of TPS1 codes (as used in tps1rc.cfg to communicate with the TPS
through the API) and corresponding names.
}
\keyword{datasets}
