% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsChangeIonMode}
\alias{TpsChangeIonMode}
\title{Changes ion mode and sets target values to 0.}
\usage{
TpsChangeIonMode(ionMode)
}
\arguments{
\item{ionMode}{0: positive ion mode, 1: negative ion mode}
}
\description{
\code{TpsChangeIonMode} changes ion mode (and sets target values to 0).
}
\details{
Note: This is an undocumented function of TofDaqDll.dll which must be used
with care. First shut the TPS down, then wait >30 s before changing the ion
mode. Not conforming to this might cause hardware damage.
}
\seealso{
Other TPS functions: \code{\link{TpsConnect2}},
  \code{\link{TpsConnect}}, \code{\link{TpsDisconnect}},
  \code{\link{TpsGetActiveFilament}},
  \code{\link{TpsGetLastSetValue}},
  \code{\link{TpsGetModuleCodes}},
  \code{\link{TpsGetModuleLimits}},
  \code{\link{TpsGetMonitorValue}},
  \code{\link{TpsGetNbrModules}},
  \code{\link{TpsGetStatus}},
  \code{\link{TpsGetTargetValue}},
  \code{\link{TpsInitialize}},
  \code{\link{TpsLoadSetFile}},
  \code{\link{TpsSaveSetFile}},
  \code{\link{TpsSetActiveFilament}},
  \code{\link{TpsSetAllVoltages}},
  \code{\link{TpsSetTargetValue}},
  \code{\link{TpsShutdown}}
}
\concept{TPS functions}
