% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsConnect2}
\alias{TpsConnect2}
\title{Connects to a local or remote TPS.}
\usage{
TpsConnect2(ip, type)
}
\arguments{
\item{ip}{TPS2 host name or IP.}

\item{type}{TPS type (0: 1st generation TPS, 1: 2nd generation TPS).}
}
\description{
\code{TpsConnect2} connects to a local or remote TPS (TPS1: type = 0,
TPS2: type = 1)
}
\seealso{
Other TPS functions: \code{\link{TpsChangeIonMode}},
  \code{\link{TpsConnect}}, \code{\link{TpsDisconnect}},
  \code{\link{TpsGetActiveFilament}},
  \code{\link{TpsGetLastSetValue}},
  \code{\link{TpsGetModuleCodes}},
  \code{\link{TpsGetModuleLimits}},
  \code{\link{TpsGetMonitorValue}},
  \code{\link{TpsGetNbrModules}},
  \code{\link{TpsGetStatus}},
  \code{\link{TpsGetTargetValue}},
  \code{\link{TpsInitialize}},
  \code{\link{TpsLoadSetFile}},
  \code{\link{TpsSaveSetFile}},
  \code{\link{TpsSetActiveFilament}},
  \code{\link{TpsSetAllVoltages}},
  \code{\link{TpsSetTargetValue}},
  \code{\link{TpsShutdown}}
}
\concept{TPS functions}
