% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsDisconnect}
\alias{TpsDisconnect}
\title{Disconnects from a remote control enabled TPSController software.}
\usage{
TpsDisconnect()
}
\description{
\code{TpsDisconnect} disconnects from a remote control enabled TPSController
software.
}
\seealso{
Other TPS functions: \code{\link{TpsChangeIonMode}},
  \code{\link{TpsConnect2}}, \code{\link{TpsConnect}},
  \code{\link{TpsGetActiveFilament}},
  \code{\link{TpsGetLastSetValue}},
  \code{\link{TpsGetModuleCodes}},
  \code{\link{TpsGetModuleLimits}},
  \code{\link{TpsGetMonitorValue}},
  \code{\link{TpsGetNbrModules}},
  \code{\link{TpsGetStatus}},
  \code{\link{TpsGetTargetValue}},
  \code{\link{TpsInitialize}},
  \code{\link{TpsLoadSetFile}},
  \code{\link{TpsSaveSetFile}},
  \code{\link{TpsSetActiveFilament}},
  \code{\link{TpsSetAllVoltages}},
  \code{\link{TpsSetTargetValue}},
  \code{\link{TpsShutdown}}
}

