% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{TpsGetActiveFilament}
\alias{TpsGetActiveFilament}
\title{Gets the currently active filament.}
\usage{
TpsGetActiveFilament()
}
\value{
Returns 0 for Filament 1, 1 for Filament 2.
}
\description{
\code{TpsGetActiveFilament} gets the currently active filament.
}
\details{
Note that \code{TpsGetMonitorValue} does not work to query the filament
number. Use this function instead.
}
\seealso{
Other TPS functions: \code{\link{TpsChangeIonMode}},
  \code{\link{TpsConnect2}}, \code{\link{TpsConnect}},
  \code{\link{TpsDisconnect}},
  \code{\link{TpsGetLastSetValue}},
  \code{\link{TpsGetModuleCodes}},
  \code{\link{TpsGetModuleLimits}},
  \code{\link{TpsGetMonitorValue}},
  \code{\link{TpsGetNbrModules}},
  \code{\link{TpsGetStatus}},
  \code{\link{TpsGetTargetValue}},
  \code{\link{TpsInitialize}},
  \code{\link{TpsLoadSetFile}},
  \code{\link{TpsSaveSetFile}},
  \code{\link{TpsSetActiveFilament}},
  \code{\link{TpsSetAllVoltages}},
  \code{\link{TpsSetTargetValue}},
  \code{\link{TpsShutdown}}
}

