% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsGetModuleProperties}
\alias{TpsGetModuleProperties}
\title{Gets capabilities and label for a given RC code.}
\usage{
TpsGetModuleProperties(moduleCode)
}
\arguments{
\item{moduleCode}{Module code.}
}
\value{
List with the properties (hasMonitor, isSettable, isTrigger) and
the label associated with \code{moduleCode} (can come from HW or cfg).
}
\description{
\code{TpsGetModuleProperties} gets capabilities and label for a given RC code.
}
\seealso{
Other TPS functions: 
\code{\link{TpsChangeIonMode}()},
\code{\link{TpsConnect2}()},
\code{\link{TpsConnect}()},
\code{\link{TpsDisconnect}()},
\code{\link{TpsGetActiveFilament}()},
\code{\link{TpsGetLastSetValue}()},
\code{\link{TpsGetModuleCodes}()},
\code{\link{TpsGetModuleLimits}()},
\code{\link{TpsGetMonitorValue}()},
\code{\link{TpsGetNbrModules}()},
\code{\link{TpsGetNmtState}()},
\code{\link{TpsGetStatus}()},
\code{\link{TpsGetTargetValue}()},
\code{\link{TpsInitialize}()},
\code{\link{TpsLoadSetFile2}()},
\code{\link{TpsLoadSetFile}()},
\code{\link{TpsSaveSetFileRc}()},
\code{\link{TpsSaveSetFile}()},
\code{\link{TpsSetActiveFilament}()},
\code{\link{TpsSetAllVoltages}()},
\code{\link{TpsSetNmtCmd}()},
\code{\link{TpsSetTargetValue}()},
\code{\link{TpsShutdown}()}
}
\concept{TPS functions}
