% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsGetMonitorValue}
\alias{TpsGetMonitorValue}
\title{Gets the last reported monitor value for a given module.}
\usage{
TpsGetMonitorValue(moduleCode)
}
\arguments{
\item{moduleCode}{Module code.}
}
\description{
\code{TpsGetMonitorValue} gets the last reported monitor value for a given
module.
}
\seealso{
Other TPS functions: 
\code{\link{TpsChangeIonMode}()},
\code{\link{TpsConnect2}()},
\code{\link{TpsConnect}()},
\code{\link{TpsDisconnect}()},
\code{\link{TpsGetActiveFilament}()},
\code{\link{TpsGetLastSetValue}()},
\code{\link{TpsGetModuleCodes}()},
\code{\link{TpsGetModuleLimits}()},
\code{\link{TpsGetModuleProperties}()},
\code{\link{TpsGetNbrModules}()},
\code{\link{TpsGetNmtState}()},
\code{\link{TpsGetStatus}()},
\code{\link{TpsGetTargetValue}()},
\code{\link{TpsInitialize}()},
\code{\link{TpsLoadSetFile2}()},
\code{\link{TpsLoadSetFile}()},
\code{\link{TpsSaveSetFileRc}()},
\code{\link{TpsSaveSetFile}()},
\code{\link{TpsSetActiveFilament}()},
\code{\link{TpsSetAllVoltages}()},
\code{\link{TpsSetNmtCmd}()},
\code{\link{TpsSetTargetValue}()},
\code{\link{TpsShutdown}()}
}
\concept{TPS functions}
