% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsGetNmtState}
\alias{TpsGetNmtState}
\title{Queries the NMT state of a CANopen node.}
\usage{
TpsGetNmtState(moduleCode)
}
\arguments{
\item{moduleCode}{Module code.}
}
\description{
\code{TpsGetNmtState} queries the NMT (Network Management) state of the
CANopen node associated with RC code \code{moduleCode}. Possible returned
NMT states are: 0 (0x00, boot up), 4 (0x04, stopped), 5 (0x05, operational)
and 127 (0x7f, pre-operational).
}
\seealso{
Other TPS functions: 
\code{\link{TpsChangeIonMode}()},
\code{\link{TpsConnect2}()},
\code{\link{TpsConnect}()},
\code{\link{TpsDisconnect}()},
\code{\link{TpsGetActiveFilament}()},
\code{\link{TpsGetLastSetValue}()},
\code{\link{TpsGetModuleCodes}()},
\code{\link{TpsGetModuleLimits}()},
\code{\link{TpsGetModuleProperties}()},
\code{\link{TpsGetMonitorValue}()},
\code{\link{TpsGetNbrModules}()},
\code{\link{TpsGetStatus}()},
\code{\link{TpsGetTargetValue}()},
\code{\link{TpsInitialize}()},
\code{\link{TpsLoadSetFile2}()},
\code{\link{TpsLoadSetFile}()},
\code{\link{TpsSaveSetFileRc}()},
\code{\link{TpsSaveSetFile}()},
\code{\link{TpsSetActiveFilament}()},
\code{\link{TpsSetAllVoltages}()},
\code{\link{TpsSetNmtCmd}()},
\code{\link{TpsSetTargetValue}()},
\code{\link{TpsShutdown}()}
}
\concept{TPS functions}
