% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsInitialize}
\alias{TpsInitialize}
\title{Initializes TPS.}
\usage{
TpsInitialize()
}
\description{
\code{TpsInitialize} initializes the TPS.
}
\seealso{
Other TPS functions: \code{\link{TpsChangeIonMode}},
  \code{\link{TpsConnect2}}, \code{\link{TpsConnect}},
  \code{\link{TpsDisconnect}},
  \code{\link{TpsGetActiveFilament}},
  \code{\link{TpsGetLastSetValue}},
  \code{\link{TpsGetModuleCodes}},
  \code{\link{TpsGetModuleLimits}},
  \code{\link{TpsGetMonitorValue}},
  \code{\link{TpsGetNbrModules}},
  \code{\link{TpsGetStatus}},
  \code{\link{TpsGetTargetValue}},
  \code{\link{TpsLoadSetFile}},
  \code{\link{TpsSaveSetFile}},
  \code{\link{TpsSetActiveFilament}},
  \code{\link{TpsSetAllVoltages}},
  \code{\link{TpsSetTargetValue}},
  \code{\link{TpsShutdown}}
}
\concept{TPS functions}
