% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsLoadSetFile}
\alias{TpsLoadSetFile}
\title{Loads a TPS set file from disk and sets all values.}
\usage{
TpsLoadSetFile(setFile)
}
\arguments{
\item{setFile}{Path/filename of the set file to load.}
}
\description{
\code{TpsLoadSetFile} loads a TPS set file from disk and sets all values.
}
\section{Warning}{

This does not just load the file (as the function name might suggest), but
also immediately sets all values.
}

\seealso{
Other TPS functions: \code{\link{TpsChangeIonMode}},
  \code{\link{TpsConnect2}}, \code{\link{TpsConnect}},
  \code{\link{TpsDisconnect}},
  \code{\link{TpsGetActiveFilament}},
  \code{\link{TpsGetLastSetValue}},
  \code{\link{TpsGetModuleCodes}},
  \code{\link{TpsGetModuleLimits}},
  \code{\link{TpsGetMonitorValue}},
  \code{\link{TpsGetNbrModules}},
  \code{\link{TpsGetStatus}},
  \code{\link{TpsGetTargetValue}},
  \code{\link{TpsInitialize}},
  \code{\link{TpsSaveSetFile}},
  \code{\link{TpsSetActiveFilament}},
  \code{\link{TpsSetAllVoltages}},
  \code{\link{TpsSetTargetValue}},
  \code{\link{TpsShutdown}}
}
