% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsLoadSetFile2}
\alias{TpsLoadSetFile2}
\title{Loads a TPS set file and sets some values.}
\usage{
TpsLoadSetFile2(setFile, blackListArray, whiteListArray)
}
\arguments{
\item{setFile}{Path/filename of the set file to load.}

\item{blackListArray}{RC code array for blacklist.}

\item{whiteListArray}{RC code array for whitelist .}
}
\description{
\code{TpsLoadSetFile2} loads a TPS set file and only sets whitelisted
values or all values except blacklisted RC codes.
}
\details{
The only 3 supported modes to call this function are:
\enumerate{
  \item \code{TpsLoadSetFile2(setFile, NULL, NULL)}, this is the same as \code{\link{TpsLoadSetFile}}
  \item \code{TpsLoadSetFile2(setFile, blackListArray, NULL)} sets the values from setFile except RC codes in blackListArray
  \item \code{TpsLoadSetFile2(setFile, NULL, whiteListArray)} sets only the values from setFile that are also in whiteListArray
}
}
\seealso{
Other TPS functions: 
\code{\link{TpsChangeIonMode}()},
\code{\link{TpsConnect2}()},
\code{\link{TpsConnect}()},
\code{\link{TpsDisconnect}()},
\code{\link{TpsGetActiveFilament}()},
\code{\link{TpsGetLastSetValue}()},
\code{\link{TpsGetModuleCodes}()},
\code{\link{TpsGetModuleLimits}()},
\code{\link{TpsGetModuleProperties}()},
\code{\link{TpsGetMonitorValue}()},
\code{\link{TpsGetNbrModules}()},
\code{\link{TpsGetNmtState}()},
\code{\link{TpsGetStatus}()},
\code{\link{TpsGetTargetValue}()},
\code{\link{TpsInitialize}()},
\code{\link{TpsLoadSetFile}()},
\code{\link{TpsSaveSetFileRc}()},
\code{\link{TpsSaveSetFile}()},
\code{\link{TpsSetActiveFilament}()},
\code{\link{TpsSetAllVoltages}()},
\code{\link{TpsSetNmtCmd}()},
\code{\link{TpsSetTargetValue}()},
\code{\link{TpsShutdown}()}
}
\concept{TPS functions}
