% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{TpsSetActiveFilament}
\alias{TpsSetActiveFilament}
\title{Sets the active filament.}
\usage{
TpsSetActiveFilament(activeFilament = 0)
}
\arguments{
\item{activeFilament}{0 for Filament 1, 1 for Filament 2.}
}
\description{
\code{TpsSetActiveFilament} sets the active filament.
}
\details{
Note that \code{TpsSetTargetValue} does not work to set the filament
number. Use this function instead.
}
\seealso{
Other TPS functions: \code{\link{TpsChangeIonMode}},
  \code{\link{TpsConnect2}}, \code{\link{TpsConnect}},
  \code{\link{TpsDisconnect}},
  \code{\link{TpsGetActiveFilament}},
  \code{\link{TpsGetLastSetValue}},
  \code{\link{TpsGetModuleCodes}},
  \code{\link{TpsGetModuleLimits}},
  \code{\link{TpsGetMonitorValue}},
  \code{\link{TpsGetNbrModules}},
  \code{\link{TpsGetStatus}},
  \code{\link{TpsGetTargetValue}},
  \code{\link{TpsInitialize}},
  \code{\link{TpsLoadSetFile}},
  \code{\link{TpsSaveSetFile}},
  \code{\link{TpsSetAllVoltages}},
  \code{\link{TpsSetTargetValue}},
  \code{\link{TpsShutdown}}
}

