% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{TpsSetAllVoltages}
\alias{TpsSetAllVoltages}
\title{Sets all voltages.}
\usage{
TpsSetAllVoltages()
}
\description{
\code{TpsSetAllVoltages} sets all voltages.
}
\seealso{
Other TPS functions: \code{\link{TpsChangeIonMode}},
  \code{\link{TpsConnect2}}, \code{\link{TpsConnect}},
  \code{\link{TpsDisconnect}},
  \code{\link{TpsGetActiveFilament}},
  \code{\link{TpsGetLastSetValue}},
  \code{\link{TpsGetModuleCodes}},
  \code{\link{TpsGetModuleLimits}},
  \code{\link{TpsGetMonitorValue}},
  \code{\link{TpsGetNbrModules}},
  \code{\link{TpsGetStatus}},
  \code{\link{TpsGetTargetValue}},
  \code{\link{TpsInitialize}},
  \code{\link{TpsLoadSetFile}},
  \code{\link{TpsSaveSetFile}},
  \code{\link{TpsSetActiveFilament}},
  \code{\link{TpsSetTargetValue}},
  \code{\link{TpsShutdown}}
}

