% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TpsSetNmtCmd}
\alias{TpsSetNmtCmd}
\title{Sets the NMT state of a CANopen node.}
\usage{
TpsSetNmtCmd(moduleCode, nmtState)
}
\arguments{
\item{moduleCode}{Module code.}

\item{nmtState}{New NMT state for node .}
}
\description{
\code{TpsSetNmtCmd} sets the NMT (Network Management) state of the
CANopen node associated with RC code \code{moduleCode}. Valid values for
nmtState are 1 (0x01, operational), 2 (0x02, stop), 128 (0x80, pre-operational),
129 (0x81, reset node) or 130 (0x82, reset communication).
}
\seealso{
Other TPS functions: 
\code{\link{TpsChangeIonMode}()},
\code{\link{TpsConnect2}()},
\code{\link{TpsConnect}()},
\code{\link{TpsDisconnect}()},
\code{\link{TpsGetActiveFilament}()},
\code{\link{TpsGetLastSetValue}()},
\code{\link{TpsGetModuleCodes}()},
\code{\link{TpsGetModuleLimits}()},
\code{\link{TpsGetModuleProperties}()},
\code{\link{TpsGetMonitorValue}()},
\code{\link{TpsGetNbrModules}()},
\code{\link{TpsGetNmtState}()},
\code{\link{TpsGetStatus}()},
\code{\link{TpsGetTargetValue}()},
\code{\link{TpsInitialize}()},
\code{\link{TpsLoadSetFile2}()},
\code{\link{TpsLoadSetFile}()},
\code{\link{TpsSaveSetFileRc}()},
\code{\link{TpsSaveSetFile}()},
\code{\link{TpsSetActiveFilament}()},
\code{\link{TpsSetAllVoltages}()},
\code{\link{TpsSetTargetValue}()},
\code{\link{TpsShutdown}()}
}
\concept{TPS functions}
