% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwH5Dll.R
\docType{data}
\name{TwRetVal}
\alias{TwRetVal}
\title{Return value.}
\format{An object of class \code{character} of length 13.}
\usage{
TwRetVal
}
\description{
Translation of the integer \code{TwRetVal} return value into a string.
}
\keyword{internal}

