% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UnregisterUserData}
\alias{UnregisterUserData}
\title{Unregisters a data source.}
\usage{
UnregisterUserData(Location)
}
\arguments{
\item{Location}{Location of group in HDF5 file identifying the user data.}
}
\description{
\code{UnregisterUserData} unregisters a data source previously registered
with \code{\link{RegisterUserDataBuf}} or \code{\link{RegisterUserDataWrite}}.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddLogEntry}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UpdateUserData}}
}

