% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{UpdateUserData}
\alias{UpdateUserData}
\title{Updates the values for a registered data source.}
\usage{
UpdateUserData(Location, NbrElements, Data)
}
\arguments{
\item{Location}{Location of group in HDF5 file where the datasets are created.}

\item{NbrElements}{Number of elements to update.}

\item{Data}{Vector of length \code{NbrElements} containing the new data.}
}
\description{
\code{UpdateUserData} updates the values for a registered data source.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}}
}

