% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{WaitForEndOfAcquisition}
\alias{WaitForEndOfAcquisition}
\title{Waits for the end of the current acquisition.}
\usage{
WaitForEndOfAcquisition(timeout)
}
\arguments{
\item{timeout}{Timeout in ms.}
}
\description{
\code{WaitForEndOfAcquisition} waits for the end of the current acquisition.
If \code{NbrRuns > 1} this function waits for the end of the last acquisition.
}

