% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaitForNewData}
\alias{WaitForNewData}
\title{Waits for new data.}
\usage{
WaitForNewData(timeout, WaitForEventReset)
}
\arguments{
\item{timeout}{Timeout in ms.}

\item{WaitForEventReset}{If \code{TRUE} (default) waits for application to
reset data available event before returning.}
}
\description{
\code{WaitForNewData} waits for new data. Returns when new data is
available or when timed out.
}

