% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaitingForDioStartSignal}
\alias{WaitingForDioStartSignal}
\title{Checks if TofDaq recorder is waiting for a start signal.}
\usage{
WaitingForDioStartSignal()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{WaitingForDioStartSignal} checks if TofDaq recorder is waiting for a
start signal.
}
\details{
Allows to query whether TofDaq recorder is currently waiting for a digital
start signal (or the SW override signal, see \code{\link{SendDioStartSignal}}). Note
that a \code{\link{StartAcquisition}} needs to be issued before this function can return
\code{TRUE}. It can take several seconds from the moment \code{\link{StartAcquisition}} is
called until this function returns \code{TRUE}.
}
