% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WriteNetCdfTimeSeriesFile}
\alias{WriteNetCdfTimeSeriesFile}
\title{Writes a ANDI chromatography file.}
\usage{
WriteNetCdfTimeSeriesFile(filename, inject_ts, expTitle, operator_name,
  company_method_name, source_file_reference, retention_unit,
  detector_unit, sample_name, raw_data_table_name, retention, ordinate)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}

\item{inject_ts}{Injection timestamp (number of 100-nanosecond intervals since
January 1, 1601 UTC, Windows FILETIME) passed as a string.}

\item{expTitle}{Experiment title.}

\item{operator_name}{Operator name.}

\item{company_method_name}{Company method name.}

\item{source_file_reference}{Source file reference.}

\item{retention_unit}{Retention unit, e.g. "[s]".}

\item{detector_unit}{Detector unit, e.g. "[cps]".}

\item{sample_name}{Sample name.}

\item{raw_data_table_name}{Dataset name.}

\item{retention}{Time axis data.}

\item{ordinate}{Intensity axis data.}
}
\description{
\code{WriteNetCdfTimeSeriesFile} writes a ANDI chromatography file.
}
