% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddLogEntry}
\alias{AddLogEntry}
\title{Adds an entry to the acquisition log.}
\usage{
AddLogEntry(LogEntryText)
}
\arguments{
\item{LogEntryText}{Log text (max. 255 characters).}
}
\description{
\code{AddLogEntry} adds an entry to the acquisition log.
}
\details{
Note: specifying the log entry time (for adding entries in the past) is not
supported.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}

