% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddLogEntry}
\alias{AddLogEntry}
\title{Adds an entry to the acquisition log.}
\usage{
AddLogEntry(LogEntryText, LogEntryTime)
}
\arguments{
\item{LogEntryText}{Log text (max. 255 characters).}

\item{LogEntryTime}{Log entry time (number of 100-nanosecond intervals since
January 1, 1601 UTC) passed as a string. Set it to "0" for "now".}
}
\description{
\code{AddLogEntry} adds an entry to the acquisition log.
}
\seealso{
Other Data storage functions: \code{\link{AddAttributeDouble}},
  \code{\link{AddAttributeInt}},
  \code{\link{AddAttributeString}},
  \code{\link{AddUserDataMultiRow}},
  \code{\link{AddUserData}},
  \code{\link{GetRegUserDataDesc}},
  \code{\link{GetRegUserDataSources}},
  \code{\link{KeepFileOpen}},
  \code{\link{QueryRegUserDataSize}},
  \code{\link{ReadRegUserData}},
  \code{\link{RegisterUserDataBuf}},
  \code{\link{RegisterUserDataNoStore}},
  \code{\link{RegisterUserDataWrite}},
  \code{\link{UnregisterUserData}},
  \code{\link{UpdateUserData}}
}

