% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AutoSetupDaqDevice}
\alias{AutoSetupDaqDevice}
\title{Auto setup routine for the DAQ device.}
\usage{
AutoSetupDaqDevice()
}
\description{
Auto setup routine for the DAQ device. Currently implemented only for
AP240 averager and ndigo5G.
}

