% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameter}
\alias{GetDaqParameter}
\title{Gets a single parameter as a string.}
\usage{
GetDaqParameter(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}
}
\description{
\code{GetDaqParameter} gets a single parameter as a string.
}
