% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDaqParameterFloat}
\alias{GetDaqParameterFloat}
\title{Gets a single float parameter.}
\usage{
GetDaqParameterFloat(Parameter)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}
}
\description{
\code{GetDaqParameterFloat} gets a single float parameter.
}
