% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetDescriptor}
\alias{GetDescriptor}
\title{Gets various information about the active acquisition.}
\usage{
GetDescriptor()
}
\value{
A list containing the TSharedMemoryDesc structure
}
\description{
\code{GetDescriptor} retrieves the current TSharedMemoryDesc structure.
TSharedMemoryDesc contains various static information about the active
acquisition that can be retrieved by \code{GetDaqParameter} functions but
also information of DAQ progress.
See \emph{/doc/TofDaqDll.htm} for more details.
}
\details{
Unsigned int parameters are returned as numeric. int64 and unsigned int64
parameters are returned as string by the dll and then converted to integer64
using \code{\link[bit64]{as.integer64}}.
}

