% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetIsotopePattern2}
\alias{GetIsotopePattern2}
\title{Calculates the isotope pattern of a molecule.}
\usage{
GetIsotopePattern2(molecule, abundanceLimit)
}
\arguments{
\item{molecule}{Molecule string.}

\item{abundanceLimit}{Absolute abundance limit for the generated pattern.}
}
\value{
List with mass and abundace vectors. Abundances are normalized
relative to the most abundant peak of the isotope pattern.
}
\description{
\code{GetIsotopePattern2} parses the molecular formula and returns the
isotope pattern (mass and abundance).
}
\details{
\code{GetIsotopePattern2} is the same as \code{GetIsotopePattern} but uses an
exact algorithm and is suitable only for small molecules.
See also \code{\link[enviPat:isopattern]{enviPat::isopattern()}}
for an alternative function to calculate accurate isotope patterns.

Note that abundances are returned relative to the most abundant peak of the isotope
pattern. The \code{abundanceLimit}, however, is an absolute abundance.

The formula parser for the molecule string understands the (case sensitive)
element labels and round and curly brackets (square brackets are reserved for
isotope specification). Numbers (multipliers) have to follow either directly
the element symbol or a closing bracket. Specific isoptopes are specified in
square brackets with the mass number before the element label (e.g. [2H],
[14C] or [235U]). Charge indicators (+, -) have to be the last characters in
the formula string, multiple charges require multiple charge symbols (e.g.
doubly charged calcium is Ca++ and not Ca2+ which is the correct syntax for a
singly charged calcium dimer).
}
\examples{
GetIsotopePattern2("CO2", 1e-5)
}
\seealso{
Other Chemistry functions: 
\code{\link{DecomposeMass2}()},
\code{\link{DecomposeMass}()},
\code{\link{GetIsotopePattern}()},
\code{\link{GetMoleculeMass}()}
}
\concept{Chemistry functions}
