% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetPeakData}
\alias{GetPeakData}
\title{Gets data stored in /PeakData/PeakData from HDF5 data file.}
\usage{
GetPeakData(
  Filename,
  peakOffset,
  peakCount,
  segOffset,
  segCount,
  bufOffset,
  bufCount,
  writeOffset,
  writeCount
)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{peakOffset}{Peak offset.}

\item{peakCount}{Peak count.}

\item{segOffset}{Segment offset.}

\item{segCount}{Segment count.}

\item{bufOffset}{Buf offset.}

\item{bufCount}{Buf count.}

\item{writeOffset}{Write offset.}

\item{writeCount}{Write count.}
}
\value{
A vector of length peakCount*segCount*bufCount*writeCount.
}
\description{
\code{GetPeakData} gets data stored in \code{/PeakData/PeakData} from HDF5 data file.
}
