% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRegUserDataFromH5}
\alias{GetRegUserDataFromH5}
\title{Reads entries from a registered data source dataset.}
\usage{
GetRegUserDataFromH5(Filename, location, bufIndex, writeIndex,
  readDescription = TRUE)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{bufIndex}{Buf index.}

\item{writeIndex}{Write index.}

\item{readDescription}{If \code{TRUE} (default) the data descripton is read,
if \code{false} the data description is not read.}
}
\value{
A list containing the registered user data and description, or a
numeric vector containing the data only if \code{readDescription = FALSE}.
}
\description{
\code{GetRegUserDataFromH5} reads an entry (or all entries) from a
registered data source dataset (created by \code{RegisterUserData...} functions).
}
\details{
If \code{bufIndex = writeIndex = -1}, all data are read.
}
\examples{
\dontrun{
GetRegUserDataFromH5("path/to/file.h5", location = )
}
}

