% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwH5Dll.R
\name{GetSumSpectrumFromH5}
\alias{GetSumSpectrumFromH5}
\title{Sum spectrum from HDF5 data file.}
\usage{
GetSumSpectrumFromH5(filename, normalize = FALSE)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}

\item{normalize}{If \code{FALSE} (default) the spectrum is reported as sum,
if \code{TRUE} the spectrum is normalized to counts per extraction.}
}
\value{
A vector containing the sum spectrum.
}
\description{
\code{GetSumSpectrumFromH5} reads the sum spectrum (or average spectrum
depending on \code{Normalize} flag) from the HDF5 file.
}
\examples{
\dontrun{
GetSumSpectrumFromH5("path/to/file.h5")
}
}

