% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{H5SetMassCalibDynamic}
\alias{H5SetMassCalibDynamic}
\title{Stores dynamic mass calibration for a given spectrum in the data file.}
\usage{
H5SetMassCalibDynamic(filename, writeIndex, par)
}
\arguments{
\item{filename}{Path/filename of the HDF5 file.}

\item{writeIndex}{Write index of the calibration to store.}

\item{par}{Numeric vector holding the actual calibration values. It
must be the same number of parameters as the global mass calibration.}
}
\description{
\code{H5SetMassCalibDynamic} stores dynamic mass calibration for a given
spectrum in the data file.
}
\details{
This function can be used to delete the dynamic calibration information by
passing the special parameter set: \code{writeIndex} = -1 and \code{par} =
\code{NULL}.
}
