% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KeepSharedMemMapped}
\alias{KeepSharedMemMapped}
\title{Keeps the shared memory acquisition buffers mapped.}
\usage{
KeepSharedMemMapped()
}
\description{
\code{KeepSharedMemMapped} Keeps the shared memory acquisition buffers mapped.
}
\details{
The DLL periodically unmaps the shared memory to give the recorder
application the possibility to (re)allocate the shared buffers. Call this
function if you want to make sure that the shared memory pointers stay
valid while you work with them. In this case you must call
\code{\link{ReleaseSharedMemory}} explicitly when finished with your
processing operation.
}

