% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ManualContinueNeeded}
\alias{ManualContinueNeeded}
\title{Indicates if the TofDaq recorder expects a continue event.}
\usage{
ManualContinueNeeded()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{ManualContinueNeeded} indicates if the TofDaq recorder expects a
continue event.
}
