% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofDaqDll.R
\name{SaveIniFile}
\alias{SaveIniFile}
\title{Saves the current configuration (*.ini) to disk.}
\usage{
SaveIniFile(IniFile)
}
\arguments{
\item{IniFile}{Path/filename of the configuration file. If \code{IniFile} is
an empty string, "TwApiTmpIni.ini" in the TofDaq recorder directory will be
used.}
}
\description{
\code{SaveIniFile} saves the current configuration (*.ini) to disk.
}

