% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetDaqParameter}
\alias{SetDaqParameter}
\title{Sets a single parameter.}
\usage{
SetDaqParameter(Parameter, ValueString)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}

\item{ValueString}{Value as a string.}
}
\description{
\code{SetDaqParameter} sets a single parameter.
}

