% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetDaqParameterFloat}
\alias{SetDaqParameterFloat}
\title{Sets a single parameter with a float value.}
\usage{
SetDaqParameterFloat(Parameter, Value)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\href{https://htmlpreview.github.io/?https://github.com/pasturm/TofDaqR/blob/master/tools/doc/TofDaqDll.htm#parameter_list}{TofDaq API documentation}
for a list of all available parameters.}

\item{Value}{Numeric value.}
}
\description{
\code{SetDaqParameterFloat} sets a single parameter with a float value.
}
