% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetDaqParameterInt}
\alias{SetDaqParameterInt}
\title{Sets a single parameter with an integer value.}
\usage{
SetDaqParameterInt(Parameter, Value)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\href{https://htmlpreview.github.io/?https://github.com/pasturm/TofDaqR/blob/master/tools/doc/TofDaqDll.htm#parameter_list}{TofDaq API documentation}
for a list of all available parameters.}

\item{Value}{Integer value.}
}
\description{
\code{SetDaqParameterInt} sets a single parameter with an integer value.
}
