% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetDaqParameterInt64}
\alias{SetDaqParameterInt64}
\title{Sets a single parameter with an int64 value.}
\usage{
SetDaqParameterInt64(Parameter, Value)
}
\arguments{
\item{Parameter}{Parameter name as a string. See
\emph{/doc/TofDaqDll.htm#parameter_list} for a list of all available parameters.}

\item{Value}{int64 value passed as a string.}
}
\description{
\code{SetDaqParameterInt64} sets a single parameter with an int64 value
(passed as a string).
}
