% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SetFloatAttributeInH5}
\alias{SetFloatAttributeInH5}
\title{Writes a float attribute to the HDF5 file.}
\usage{
SetFloatAttributeInH5(Filename, location, name, attribute)
}
\arguments{
\item{Filename}{Path/filename of the HDF5 file.}

\item{location}{Location of the group or dataset where the attribute is attached to.}

\item{name}{Attribute name.}

\item{attribute}{Float attribute.}
}
\description{
\code{SetFloatAttributeInH5} writes a float attribute to the HDF5 file.
}
\examples{
\dontrun{
SetFloatAttributeInH5("path/to/file.h5", location = , name = , attribute = )
}
}

