% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwToolDll.R
\name{SiProcessSpectrum}
\alias{SiProcessSpectrum}
\title{Processes a spectrum.}
\usage{
SiProcessSpectrum(spectrum, specType = 0)
}
\arguments{
\item{spectrum}{Vector holding the spectrum to process.}

\item{specType}{Spectrum type index (non-negative integer).}
}
\value{
A list with the baseline and threshold value.
}
\description{
\code{SiProcessSpectrum} processes a spectrum according to the options set for
it's spectrum type.
}
\details{
See \code{\link{SiProcessSpectrumFromShMem}} for another variant of this
function, where the spectrum is directly read from shared memory.
}
\seealso{
Other Single ion histogramming functions.: \code{\link{SiCleanup}},
  \code{\link{SiEvalPhd}}, \code{\link{SiFitPhd}},
  \code{\link{SiFitRateFromPhd}},
  \code{\link{SiGetHistogram}},
  \code{\link{SiGetSumHistogram}},
  \code{\link{SiInitializeHistograms}},
  \code{\link{SiProcessSpectrumFromShMem}},
  \code{\link{SiResetHistograms}},
  \code{\link{SiSetProcessingOptions}}
}

