% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwToolDll.R
\name{Tof2Mass}
\alias{Tof2Mass}
\title{Converts from sample index to mass/charge.}
\usage{
Tof2Mass(tofSample, massCalibMode = 0, p)
}
\arguments{
\item{tofSample}{Vector of sample indices to convert.}

\item{massCalibMode}{Mass calibration function to use. See below.}

\item{p}{Vector containing the calibration parameters (number depends on
\code{MassCalibMode}, see below).}
}
\value{
Mass/charge values.
}
\description{
\code{Tof2Mass} converts from sample index to mass/charge.
}
\details{
\tabular{cl}{
massCalibMode \tab Mass calibration function \cr
0 \tab \eqn{i = p_1 \sqrt(m) + p_2} \cr
1 \tab \eqn{i = p_1/\sqrt(m) + p_2} \cr
2 \tab \eqn{i = p_1 m^{p_3} + p_2} \cr
3 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 (m - p_4)^2} \cr
4 \tab \eqn{i = p_1 \sqrt(m) + p_2 + p_3 m^2 + p_4 m + p_5} \cr
5 \tab \eqn{m = p_1 i^2 + p_2 i + p_3}
}
}
\examples{
Tof2Mass(100000, massCalibMode = 0, p = c(3,5))
}
\seealso{
\code{\link{Mass2Tof}}
}

